refactor(core): split native.rs god module into focused submodules
Breaks crates/zeph-core/src/agent/tool_execution/native.rs (5084 lines,
108 functions) into 6 concern-specific files within the same module:

- llm_dispatch.rs   — LLM call orchestration and streaming
- tool_result.rs    — result classification and processing
- tier_loop.rs      — tiered parallel execution and post-dispatch phases
- metrics_compact.rs — metrics recording and server compaction
- layer_hooks.rs    — RuntimeLayer before/after hooks

native.rs becomes a thin entry point (~1100 lines) containing only the
module declarations, focus/compress tool handlers, and #[cfg(test)]
infrastructure.

All impl Agent<C> methods remain on the same type; no public API changes.
Four methods bumped to pub(super) for cross-module access within the
tool_execution module. All 8934 unit tests pass.

Closes #3664
