+++ crates/zeph-core/src/agent/tool_execution/native.rs
@@ -1,129 +1,25 @@
 // SPDX-FileCopyrightText: 2026 Andrei G <bug-ops>
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-use futures::FutureExt as _;
 use zeph_common::text::estimate_tokens;
-use zeph_llm::provider::{
-    ChatResponse, LlmProvider, Message, MessageMetadata, MessagePart, Role, ThinkingBlock,
-    ToolDefinition,
-};
-
-use super::{
-    AnomalyOutcome, retry_backoff_ms, strip_tafc_fields, tool_args_hash,
-    tool_def_to_definition_with_tafc,
-};
-use crate::agent::Agent;
-use crate::channel::{Channel, StopHint, ToolOutputEvent, ToolStartEvent};
-use crate::overflow_tools::OverflowToolExecutor;
-use tracing::Instrument;
+#[cfg(test)]
+use zeph_llm::provider::LlmProvider;
 use zeph_llm::provider::MAX_TOKENS_TRUNCATION_MARKER;
-use zeph_sanitizer::{ContentSource, ContentSourceKind};
-use zeph_skills::evolution::FailureKind;
-use zeph_tools::ExecutionContext;
-use zeph_tools::executor::ToolCall;
+use zeph_llm::provider::{ChatResponse, Message, MessagePart, Role, ToolDefinition};
 
-type ToolExecFut = futures::future::BoxFuture<
-    'static,
-    Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>,
->;
-
-/// Output produced by the tier execution loop (Phase 1).
-///
-/// `None` signals that the loop was cancelled by the user; `Some` carries the
-/// collected tool results and deferred data that the results phase needs.
-type TierLoopOutput = Option<TierLoopData>;
-
-struct TierLoopData {
-    tool_results: Vec<Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>>,
-    pending_focus_checkpoint: Option<zeph_llm::provider::Message>,
-    pending_system_hints: Vec<String>,
-}
-
-/// Per-call computed context produced by `Agent::prepare_tool_dispatch`.
-///
-/// Bundles the gating results (pre-exec block, utility action, repeat/quota/cache checks)
-/// alongside the canonical call list so the tier execution loop has a single, coherent view.
-struct ToolDispatchContext {
-    calls: Vec<ToolCall>,
-    tool_call_ids: Vec<String>,
-    tool_started_ats: Vec<std::time::Instant>,
-    pre_exec_blocked: Vec<bool>,
-    utility_actions: Vec<zeph_tools::UtilityAction>,
-    quota_blocked: bool,
-    args_hashes: Vec<u64>,
-    repeat_blocked: Vec<bool>,
-    cache_hits: Vec<Option<zeph_tools::ToolOutput>>,
-}
-
-/// Output of `Agent::classify_tool_result`: all fields derived from the raw `ToolResult`.
-///
-/// Returned by value so `process_one_tool_result` can unpack it without borrow issues.
-struct ToolResultClassification {
-    output: String,
-    is_error: bool,
-    diff: Option<zeph_tools::DiffData>,
-    inline_stats: Option<String>,
-    kept_lines: Option<Vec<usize>>,
-    locations: Option<Vec<String>>,
-    anomaly_outcome: AnomalyOutcome,
-    is_quality_failure: bool,
-    tool_err_category: Option<zeph_tools::error_taxonomy::ToolErrorCategory>,
-}
+use super::tool_def_to_definition_with_tafc;
+use crate::agent::Agent;
+use crate::channel::{Channel, StopHint};
+#[cfg(test)]
+use tracing::Instrument;
 
 impl<C: Channel> Agent<C> {
-    pub(super) async fn call_chat_with_tools_retry(
-        &mut self,
-        tool_defs: &[ToolDefinition],
-        max_attempts: usize,
-    ) -> Result<Option<ChatResponse>, crate::agent::error::AgentError> {
-        for attempt in 0..max_attempts {
-            match self.call_chat_with_tools(tool_defs).await {
-                Ok(result) => return Ok(result),
-                Err(e) if e.is_context_length_error() && attempt + 1 < max_attempts => {
-                    tracing::warn!(
-                        attempt,
-                        "chat_with_tools context length exceeded, compacting and retrying"
-                    );
-                    let _ = self
-                        .channel
-                        .send_status("context too long, compacting...")
-                        .await;
-                    let _ = self.compact_context().await?;
-                    let _ = self.channel.send_status("").await;
-                }
-                Err(e) if e.is_beta_header_rejected() && attempt + 1 < max_attempts => {
-                    // SEC-COMPACT-03: the compact-2026-01-12 beta header was rejected by the API.
-                    // The provider already set its internal flag; disable client-side gate and
-                    // retry so this turn is not lost.
-                    tracing::warn!(
-                        attempt,
-                        "server compaction beta header rejected; \
-                        falling back to client-side compaction and retrying"
-                    );
-                    self.runtime.providers.server_compaction_active = false;
-                    let _ = self
-                        .channel
-                        .send_status(
-                            "server compaction unavailable, falling back to client-side...",
-                        )
-                        .await;
-                    let _ = self.channel.send_status("").await;
-                }
-                Err(e) => return Err(e),
-            }
-        }
-        unreachable!("loop covers all attempts")
-    }
-
-    #[cfg_attr(
-        feature = "profiling",
-        tracing::instrument(name = "agent.process_response", skip_all)
-    )]
+    #[tracing::instrument(name = "core.tool.process_response", skip_all, level = "debug", err)]
     pub(crate) async fn process_response(&mut self) -> Result<(), crate::agent::error::AgentError> {
         self.services.security.flagged_urls.clear();
         self.process_response_native_tools().await
     }
-
+    #[tracing::instrument(name = "core.tool.native_loop", skip_all, level = "debug", err)]
     pub(super) async fn process_response_native_tools(
         &mut self,
     ) -> Result<(), crate::agent::error::AgentError> {
@@ -747,6 +643,13 @@ impl<C: Channel> Agent<C> {
     /// Execute one turn of the native tool loop. Returns `Ok(Some(()))` when the LLM produced
     /// a terminal text response (caller should return `Ok(())`), `Ok(None)` to continue the
     /// loop, or `Err` on a hard error.
+    #[tracing::instrument(
+        name = "core.tool.single_turn",
+        skip_all,
+        level = "debug",
+        fields(iteration),
+        err
+    )]
     async fn process_single_native_turn(
         &mut self,
         tool_defs: &[ToolDefinition],
@@ -848,2764 +751,6 @@ impl<C: Channel> Agent<C> {
         Ok(None)
     }
 
-    async fn call_chat_with_tools(
-        &mut self,
-        tool_defs: &[ToolDefinition],
-    ) -> Result<Option<ChatResponse>, crate::agent::error::AgentError> {
-        if let Some(ref tracker) = self.runtime.metrics.cost_tracker
-            && let Err(e) = tracker.check_budget()
-        {
-            self.channel
-                .send(&format!("Budget limit reached: {e}"))
-                .await?;
-            return Ok(None);
-        }
-
-        tracing::debug!(
-            tool_count = tool_defs.len(),
-            provider_name = self.provider.name(),
-            "call_chat_with_tools"
-        );
-        let llm_timeout = std::time::Duration::from_secs(self.runtime.config.timeouts.llm_seconds);
-        let start = std::time::Instant::now();
-
-        let memcot_state_for_dump =
-            match self.services.memory.extraction.memcot_accumulator.as_ref() {
-                Some(acc) => acc.current_state().await,
-                None => None,
-            };
-        let dump_id = self.prepare_chat_debug_dump(tool_defs, memcot_state_for_dump.as_deref());
-
-        // RuntimeLayer before_chat hooks (MVP: empty vec = zero iterations).
-        if let Some(sc) = self.run_before_chat_layers(tool_defs).await? {
-            return Ok(Some(sc));
-        }
-
-        // Inject accumulated LSP notes (hover, diagnostics) as a Role::System message
-        // immediately before the LLM call. At this point all tool results from the previous
-        // iteration are committed to history and there is no pending ToolUse/ToolResult pair,
-        // so inserting a System message is safe for all providers (OpenAI, Claude, Ollama).
-        // Stale notes from a prior call_chat_with_tools invocation are removed first so they
-        // never accumulate; Role::System is skipped by tool-pair summarization.
-        if self.services.session.lsp_hooks.is_some() {
-            self.remove_lsp_messages();
-            let tc = std::sync::Arc::clone(&self.runtime.metrics.token_counter);
-            if let Some(ref mut lsp) = self.services.session.lsp_hooks
-                && let Some(note_text) = lsp.drain_notes(&tc)
-            {
-                self.push_message(zeph_llm::provider::Message::from_legacy(
-                    zeph_llm::provider::Role::System,
-                    &note_text,
-                ));
-                self.recompute_prompt_tokens();
-            }
-        }
-
-        // CR-01: open LLM span before the call.
-        let trace_guard = self.runtime.debug.trace_collector.as_ref().and_then(|tc| {
-            self.runtime
-                .debug
-                .current_iteration_span_id
-                .map(|id| tc.begin_llm_request(id))
-        });
-
-        let llm_span = tracing::info_span!(
-            "llm.turn_call",
-            model = %self.runtime.config.model_name,
-            provider = self.provider.name(),
-        );
-
-        let Some(result) = self
-            .dispatch_chat_with_tools(tool_defs, llm_timeout, llm_span)
-            .await?
-        else {
-            return Ok(None);
-        };
-
-        self.record_chat_metrics_and_compact(start, &result).await?;
-
-        // Accumulate LLM chat latency into the per-turn timing accumulator (#2820).
-        let llm_ms = u64::try_from(start.elapsed().as_millis()).unwrap_or(u64::MAX);
-        self.runtime.metrics.pending_timings.llm_chat_ms = self
-            .runtime
-            .metrics
-            .pending_timings
-            .llm_chat_ms
-            .saturating_add(llm_ms);
-
-        // CR-01: close LLM span after the call completes.
-        self.record_llm_trace_span_close(trace_guard, start);
-
-        self.runtime.debug.write_chat_debug_dump(
-            dump_id,
-            &result,
-            &self.services.security.pii_filter,
-        );
-
-        // RuntimeLayer after_chat hooks (MVP: empty vec = zero iterations).
-        self.run_after_chat_layers(&result).await;
-
-        Ok(Some(result))
-    }
-
-    /// Route the LLM call through the speculative SSE path or the normal blocking path.
-    ///
-    /// When `SpeculationMode` is `Decoding` or `Both`, attempts the streaming tool path and
-    /// falls back to non-streaming if the provider does not support it or the stream fails.
-    async fn dispatch_chat_with_tools(
-        &mut self,
-        tool_defs: &[ToolDefinition],
-        llm_timeout: std::time::Duration,
-        llm_span: tracing::Span,
-    ) -> Result<Option<ChatResponse>, crate::agent::error::AgentError> {
-        let use_speculative_stream = self.services.speculation_engine.as_ref().is_some_and(|e| {
-            matches!(
-                e.mode(),
-                zeph_config::tools::SpeculationMode::Decoding
-                    | zeph_config::tools::SpeculationMode::Both
-            )
-        });
-
-        if use_speculative_stream
-            && let Ok(stream) = self
-                .provider
-                .chat_with_tools_stream(&self.msg.messages, tool_defs)
-                .await
-        {
-            let engine =
-                std::sync::Arc::clone(self.services.speculation_engine.as_ref().expect(
-                    "invariant: speculation_engine is Some (checked via is_some_and on L961)",
-                ));
-            let threshold = engine.confidence_threshold();
-            let drainer = crate::agent::speculative::stream_drainer::SpeculativeStreamDrainer::new(
-                stream, engine, threshold,
-            );
-            let drain_fut = tokio::time::timeout(llm_timeout, drainer.drive().instrument(llm_span));
-            let timeout_result = tokio::select! {
-                r = drain_fut => r,
-                () = self.runtime.lifecycle.cancel_token.cancelled() => {
-                    tracing::info!("chat_with_tools (streaming) cancelled by user");
-                    self.update_metrics(|m| m.cancellations += 1);
-                    self.channel.send("[Cancelled]").await?;
-                    return Ok(None);
-                }
-            };
-            return match timeout_result {
-                Ok(Ok(resp)) => Ok(Some(resp)),
-                Ok(Err(e)) => {
-                    tracing::warn!(error = %e, "speculative SSE stream failed, falling back");
-                    self.call_non_streaming(tool_defs, llm_timeout).await
-                }
-                Err(_) => {
-                    self.channel
-                        .send("LLM request timed out. Please try again.")
-                        .await?;
-                    Ok(None)
-                }
-            };
-        }
-        // Provider does not support tool streaming or speculative mode is off — normal path.
-        self.call_non_streaming_with_span(tool_defs, llm_timeout, llm_span)
-            .await
-    }
-
-    async fn call_non_streaming(
-        &mut self,
-        tool_defs: &[ToolDefinition],
-        llm_timeout: std::time::Duration,
-    ) -> Result<Option<ChatResponse>, crate::agent::error::AgentError> {
-        let chat_fut = tokio::time::timeout(
-            llm_timeout,
-            self.provider.chat_with_tools(&self.msg.messages, tool_defs),
-        );
-        let timeout_result = tokio::select! {
-            r = chat_fut => r,
-            () = self.runtime.lifecycle.cancel_token.cancelled() => {
-                tracing::info!("chat_with_tools cancelled by user");
-                self.update_metrics(|m| m.cancellations += 1);
-                self.channel.send("[Cancelled]").await?;
-                return Ok(None);
-            }
-        };
-        match timeout_result {
-            Ok(Ok(r)) => Ok(Some(r)),
-            Ok(Err(e)) => Err(e.into()),
-            Err(_) => {
-                self.channel
-                    .send("LLM request timed out. Please try again.")
-                    .await?;
-                Ok(None)
-            }
-        }
-    }
-
-    async fn call_non_streaming_with_span(
-        &mut self,
-        tool_defs: &[ToolDefinition],
-        llm_timeout: std::time::Duration,
-        llm_span: tracing::Span,
-    ) -> Result<Option<ChatResponse>, crate::agent::error::AgentError> {
-        let chat_fut = tokio::time::timeout(
-            llm_timeout,
-            self.provider
-                .chat_with_tools(&self.msg.messages, tool_defs)
-                .instrument(llm_span),
-        );
-        let timeout_result = tokio::select! {
-            r = chat_fut => r,
-            () = self.runtime.lifecycle.cancel_token.cancelled() => {
-                tracing::info!("chat_with_tools cancelled by user");
-                self.update_metrics(|m| m.cancellations += 1);
-                self.channel.send("[Cancelled]").await?;
-                return Ok(None);
-            }
-        };
-        match timeout_result {
-            Ok(Ok(r)) => Ok(Some(r)),
-            Ok(Err(e)) => Err(e.into()),
-            Err(_) => {
-                self.channel
-                    .send("LLM request timed out. Please try again.")
-                    .await?;
-                Ok(None)
-            }
-        }
-    }
-
-    /// Prepare and dump the request debug payload, returning the dump ID for the response dump.
-    fn prepare_chat_debug_dump(
-        &self,
-        tool_defs: &[ToolDefinition],
-        memcot_state: Option<&str>,
-    ) -> Option<u32> {
-        self.runtime
-            .debug
-            .debug_dumper
-            .as_ref()
-            .map(|d: &crate::debug_dump::DebugDumper| {
-                // Skip expensive serialization when Trace format returns early without using it.
-                let provider_request = if d.is_trace_format() {
-                    serde_json::Value::Null
-                } else {
-                    self.provider
-                        .debug_request_json(&self.msg.messages, tool_defs, false) // lgtm[rust/cleartext-logging]
-                };
-                d.dump_request(&crate::debug_dump::RequestDebugDump {
-                    model_name: &self.runtime.config.model_name,
-                    messages: &self.msg.messages,
-                    tools: tool_defs,
-                    provider_request,
-                    memcot_state,
-                })
-            })
-    }
-
-    /// Run `RuntimeLayer::before_chat` hooks. Returns `Ok(Some(sc))` when a layer short-circuits
-    /// the LLM call, `Ok(None)` when all hooks pass through.
-    async fn run_before_chat_layers(
-        &self,
-        tool_defs: &[ToolDefinition],
-    ) -> Result<Option<ChatResponse>, crate::agent::error::AgentError> {
-        if self.runtime.config.layers.is_empty() {
-            return Ok(None);
-        }
-        let conv_id_str = self
-            .services
-            .memory
-            .persistence
-            .conversation_id
-            .map(|id| id.0.to_string());
-        let ctx = crate::runtime_layer::LayerContext {
-            conversation_id: conv_id_str.as_deref(),
-            turn_number: u32::try_from(self.services.sidequest.turn_counter).unwrap_or(u32::MAX),
-        };
-        for layer in &self.runtime.config.layers {
-            let hook_result = std::panic::AssertUnwindSafe(layer.before_chat(
-                &ctx,
-                &self.msg.messages,
-                tool_defs,
-            ))
-            .catch_unwind()
-            .await;
-            match hook_result {
-                Ok(Some(sc)) => {
-                    tracing::debug!("RuntimeLayer short-circuited LLM call");
-                    return Ok(Some(sc));
-                }
-                Ok(None) => {}
-                Err(_) => tracing::warn!("RuntimeLayer::before_chat panicked, continuing"),
-            }
-        }
-        Ok(None)
-    }
-
-    /// Run `RuntimeLayer::after_chat` hooks. Panics in hooks are logged and swallowed.
-    async fn run_after_chat_layers(&self, result: &ChatResponse) {
-        if self.runtime.config.layers.is_empty() {
-            return;
-        }
-        let conv_id_str = self
-            .services
-            .memory
-            .persistence
-            .conversation_id
-            .map(|id| id.0.to_string());
-        let ctx = crate::runtime_layer::LayerContext {
-            conversation_id: conv_id_str.as_deref(),
-            turn_number: u32::try_from(self.services.sidequest.turn_counter).unwrap_or(u32::MAX),
-        };
-        for layer in &self.runtime.config.layers {
-            let hook_result = std::panic::AssertUnwindSafe(layer.after_chat(&ctx, result))
-                .catch_unwind()
-                .await;
-            if hook_result.is_err() {
-                tracing::warn!("RuntimeLayer::after_chat panicked, continuing");
-            }
-        }
-    }
-
-    /// Close the CR-01 LLM trace span after the chat call completes.
-    fn record_llm_trace_span_close(
-        &mut self,
-        guard: Option<crate::debug_dump::trace::SpanGuard>,
-        start: std::time::Instant,
-    ) {
-        if let Some(guard) = guard
-            && let Some(ref mut tc) = self.runtime.debug.trace_collector
-        {
-            let latency = u64::try_from(start.elapsed().as_millis()).unwrap_or(u64::MAX);
-            let (prompt_tokens, completion_tokens) = self.provider.last_usage().unwrap_or((0, 0));
-            tc.end_llm_request(
-                guard,
-                &crate::debug_dump::trace::LlmAttributes {
-                    model: self.runtime.config.model_name.clone(),
-                    prompt_tokens,
-                    completion_tokens,
-                    latency_ms: latency,
-                    streaming: false,
-                    cache_hit: false,
-                },
-            );
-        }
-    }
-
-    async fn record_chat_metrics_and_compact(
-        &mut self,
-        start: std::time::Instant,
-        result: &ChatResponse,
-    ) -> Result<(), crate::agent::error::AgentError> {
-        let elapsed = start.elapsed();
-        let latency = u64::try_from(elapsed.as_millis()).unwrap_or(u64::MAX);
-        let prompt_estimate = self.runtime.providers.cached_prompt_tokens;
-        let completion_heuristic = match result {
-            ChatResponse::Text(t) => estimate_tokens(t) as u64,
-            ChatResponse::ToolUse {
-                text, tool_calls, ..
-            } => {
-                let text_tokens = estimate_tokens(text.as_deref().unwrap_or(""));
-                let calls_tokens: usize = tool_calls
-                    .iter()
-                    .map(|c| {
-                        estimate_tokens(c.name.as_str()) + estimate_tokens(&c.input.to_string())
-                    })
-                    .sum();
-                (text_tokens + calls_tokens) as u64
-            }
-        };
-        let (final_prompt, final_completion) = self
-            .provider
-            .last_usage()
-            .unwrap_or((prompt_estimate, completion_heuristic));
-        let router_stats = self.provider.router_thompson_stats();
-        self.update_metrics(|m| {
-            m.api_calls += 1;
-            m.last_llm_latency_ms = latency;
-            m.context_tokens = final_prompt;
-            m.prompt_tokens += final_prompt;
-            m.completion_tokens += final_completion;
-            m.total_tokens = m.prompt_tokens + m.completion_tokens;
-            if !router_stats.is_empty() {
-                m.router_thompson_stats = router_stats;
-            }
-        });
-        // Track per-turn LLM call count for the notification gate.
-        self.runtime.lifecycle.turn_llm_requests =
-            self.runtime.lifecycle.turn_llm_requests.saturating_add(1);
-        self.record_cost_and_cache(final_prompt, final_completion);
-        self.record_successful_task();
-
-        if let Some(ref recorder) = self.runtime.metrics.histogram_recorder {
-            recorder.observe_llm_latency(elapsed);
-        }
-
-        if let Some((input_tokens, output_tokens)) = self.provider.last_usage() {
-            let context_window =
-                u64::try_from(self.provider.context_window().unwrap_or(0)).unwrap_or(0);
-            let _ = self
-                .channel
-                .send_usage(input_tokens, output_tokens, context_window)
-                .await;
-        }
-
-        // C2: server-side compaction — prune old messages and insert synthetic compaction message.
-        if let Some(raw_summary) = self.provider.take_compaction_summary() {
-            let _ = self
-                .channel
-                .send_status("Compacting context (server-side)...")
-                .await;
-            tracing::info!(
-                summary_len = raw_summary.len(),
-                messages_before = self.msg.messages.len(),
-                "server-side compaction received; pruning old messages"
-            );
-            // SEC-COMPACT-01: sanitize (McpResponse = ExternalUntrusted).
-            let source = ContentSource::new(ContentSourceKind::McpResponse);
-            let sanitized = self
-                .services
-                .security
-                .sanitizer
-                .sanitize(&raw_summary, source);
-            let summary = sanitized.body;
-            let last_user = self
-                .msg
-                .messages
-                .iter()
-                .rposition(|m| m.role == Role::User)
-                .unwrap_or(0);
-            let tail: Vec<Message> = self.msg.messages.drain(last_user..).collect();
-            self.msg.messages.clear();
-            self.msg.messages.push(Message {
-                role: Role::Assistant,
-                content: summary.clone(),
-                parts: vec![MessagePart::Compaction {
-                    summary: summary.clone(),
-                }],
-                metadata: MessageMetadata::default(),
-            });
-            self.msg.messages.extend(tail);
-            self.update_metrics(|m| m.server_compaction_events += 1);
-            let _ = self.channel.send_status("").await;
-        }
-
-        Ok(())
-    }
-
-    /// Prepend thinking blocks to the last assistant message in the context as `MessagePart`s.
-    ///
-    /// The Claude API requires `thinking`/`redacted_thinking` blocks to be preserved verbatim
-    /// in the assistant message when tool results are sent back in multi-turn conversations.
-    fn preserve_thinking_blocks(&mut self, blocks: Vec<ThinkingBlock>) {
-        if blocks.is_empty() {
-            return;
-        }
-        if let Some(last) = self.msg.messages.last_mut()
-            && last.role == Role::Assistant
-        {
-            let mut thinking_parts: Vec<MessagePart> = blocks
-                .into_iter()
-                .map(|b| match b {
-                    ThinkingBlock::Thinking {
-                        thinking,
-                        signature,
-                    } => MessagePart::ThinkingBlock {
-                        thinking,
-                        signature,
-                    },
-                    ThinkingBlock::Redacted { data } => MessagePart::RedactedThinkingBlock { data },
-                })
-                .collect();
-            // Thinking blocks must appear before text/tool_use in the assistant message.
-            thinking_parts.append(&mut last.parts);
-            last.parts = thinking_parts;
-            last.rebuild_content();
-        }
-    }
-
-    /// Handle phases 2a / 2 / 3 of the tool dispatch cycle after the tier join completes.
-    ///
-    /// - **Phase 2a**: Interactive confirmation prompts for `ConfirmationRequired` results.
-    /// - **Phase 2**: Sequential retry with exponential backoff for transient errors on retryable
-    ///   executors. Shell/non-idempotent tools keep their error result unchanged.
-    /// - **Phase 3**: Parameter-reformat placeholder path (future LLM-based reformat).
-    ///
-    /// Each phase checks the cancellation token and returns `Ok(())` on early cancel, persisting
-    /// tombstone `ToolResult` entries to satisfy the `TOOL_RESULT_GUARANTEE` invariant.
-    /// Phases 2a / 2 / 3 of the tool dispatch cycle (confirmation, retry, reformat).
-    async fn run_post_dispatch_phases(
-        &mut self,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-        calls: &[ToolCall],
-        tool_results: &mut [Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>],
-        max_retries: usize,
-        cancel: &tokio_util::sync::CancellationToken,
-    ) -> Result<(), crate::agent::error::AgentError> {
-        self.handle_confirmation_phase(tool_calls, calls, tool_results, cancel)
-            .await?;
-        self.handle_retry_phase(tool_calls, calls, tool_results, max_retries, cancel)
-            .await?;
-        self.handle_reformat_phase(tool_calls, tool_results, cancel)
-            .await?;
-        Ok(())
-    }
-
-    /// Phase 2a: resolve `ConfirmationRequired` results via interactive channel prompt.
-    async fn handle_confirmation_phase(
-        &mut self,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-        calls: &[ToolCall],
-        tool_results: &mut [Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>],
-        cancel: &tokio_util::sync::CancellationToken,
-    ) -> Result<(), crate::agent::error::AgentError> {
-        for idx in 0..tool_results.len() {
-            if cancel.is_cancelled() {
-                self.tool_executor.set_skill_env(None);
-                tracing::info!("tool execution cancelled by user");
-                self.update_metrics(|m| m.cancellations += 1);
-                self.channel.send("[Cancelled]").await?;
-                self.persist_cancelled_tool_results(tool_calls).await;
-                return Ok(());
-            }
-            let new_result =
-                if let Err(zeph_tools::ToolError::ConfirmationRequired { ref command }) =
-                    tool_results[idx]
-                {
-                    let tc = &tool_calls[idx];
-                    let prompt = if command.is_empty() {
-                        format!("Allow tool: {}?", tc.name)
-                    } else {
-                        format!("Allow command: {command}?")
-                    };
-                    Some(if self.channel.confirm(&prompt).await? {
-                        // execute_tool_call_confirmed_erased bypasses check_trust; a second
-                        // ConfirmationRequired here indicates a misconfigured executor stack.
-                        self.tool_executor
-                            .execute_tool_call_confirmed_erased(&calls[idx])
-                            .await
-                    } else {
-                        Ok(Some(zeph_tools::ToolOutput {
-                            tool_name: tc.name.clone(),
-                            summary: "[cancelled by user]".to_owned(),
-                            blocks_executed: 0,
-                            filter_stats: None,
-                            diff: None,
-                            streamed: false,
-                            terminal_id: None,
-                            locations: None,
-                            raw_response: None,
-                            claim_source: None,
-                        }))
-                    })
-                } else {
-                    None
-                };
-            if let Some(result) = new_result {
-                if let Err(ref e) = result
-                    && let Some(ref d) = self.runtime.debug.debug_dumper
-                {
-                    d.dump_tool_error(tool_calls[idx].name.as_str(), e);
-                }
-                tool_results[idx] = result;
-            }
-        }
-        Ok(())
-    }
-
-    /// Phase 2: sequential retry with exponential backoff for transient errors on retryable tools.
-    async fn handle_retry_phase(
-        &mut self,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-        calls: &[ToolCall],
-        tool_results: &mut [Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>],
-        max_retries: usize,
-        cancel: &tokio_util::sync::CancellationToken,
-    ) -> Result<(), crate::agent::error::AgentError> {
-        if max_retries == 0 {
-            return Ok(());
-        }
-        let max_retry_duration_secs = self.tool_orchestrator.max_retry_duration_secs;
-        let retry_base_ms = self.tool_orchestrator.retry_base_ms;
-        let retry_max_ms = self.tool_orchestrator.retry_max_ms;
-        for idx in 0..tool_results.len() {
-            if cancel.is_cancelled() {
-                self.tool_executor.set_skill_env(None);
-                tracing::info!("tool execution cancelled by user");
-                self.update_metrics(|m| m.cancellations += 1);
-                self.channel.send("[Cancelled]").await?;
-                self.persist_cancelled_tool_results(tool_calls).await;
-                return Ok(());
-            }
-            let is_transient = matches!(
-                tool_results[idx],
-                Err(ref e) if e.kind() == zeph_tools::ErrorKind::Transient
-            );
-            if !is_transient {
-                continue;
-            }
-            let tc = &tool_calls[idx];
-            if !self
-                .tool_executor
-                .is_tool_retryable_erased(tc.name.as_str())
-            {
-                continue;
-            }
-            let call = &calls[idx];
-            let mut attempt = 0_usize;
-            let retry_start = std::time::Instant::now();
-            let result = loop {
-                let exec_result = tokio::select! {
-                    r = self.tool_executor.execute_tool_call_erased(call).instrument(
-                        tracing::info_span!("tool_exec_retry", tool_name = %tc.name, idx = %tc.id)
-                    ) => r,
-                    () = cancel.cancelled() => {
-                        self.tool_executor.set_skill_env(None);
-                        tracing::info!("tool retry cancelled by user");
-                        self.update_metrics(|m| m.cancellations += 1);
-                        self.channel.send("[Cancelled]").await?;
-                        self.persist_cancelled_tool_results(tool_calls).await;
-                        return Ok(());
-                    }
-                };
-                match exec_result {
-                    Err(ref e)
-                        if e.kind() == zeph_tools::ErrorKind::Transient
-                            && attempt < max_retries =>
-                    {
-                        let elapsed_secs = retry_start.elapsed().as_secs();
-                        if max_retry_duration_secs > 0 && elapsed_secs >= max_retry_duration_secs {
-                            tracing::warn!(
-                                tool = %tc.name, elapsed_secs, max_retry_duration_secs,
-                                "tool retry budget exceeded, aborting retries"
-                            );
-                            break exec_result;
-                        }
-                        attempt += 1;
-                        let delay_ms = retry_backoff_ms(attempt - 1, retry_base_ms, retry_max_ms);
-                        tracing::warn!(
-                            tool = %tc.name, attempt, delay_ms, error = %e,
-                            "transient tool error, retrying with backoff"
-                        );
-                        let _ = self
-                            .channel
-                            .send_status(&format!("Retrying {}...", tc.name))
-                            .await;
-                        // Interruptible backoff sleep: cancelled if agent shuts down.
-                        tokio::select! {
-                            () = tokio::time::sleep(std::time::Duration::from_millis(delay_ms)) => {}
-                            () = cancel.cancelled() => {
-                                self.tool_executor.set_skill_env(None);
-                                tracing::info!("retry backoff interrupted by cancellation");
-                                self.update_metrics(|m| m.cancellations += 1);
-                                self.channel.send("[Cancelled]").await?;
-                                return Ok(());
-                            }
-                        }
-                        let _ = self.channel.send_status("").await;
-                        // NOTE: retry re-executions are NOT recorded in repeat-detection (CRIT-3).
-                    }
-                    result => break result,
-                }
-            };
-            tool_results[idx] = result;
-        }
-        Ok(())
-    }
-
-    /// Phase 3: parameter-reformat placeholder path (future LLM-based reformat).
-    ///
-    /// Currently logs and skips; original error is propagated unchanged to the LLM.
-    async fn handle_reformat_phase(
-        &mut self,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-        tool_results: &mut [Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>],
-        cancel: &tokio_util::sync::CancellationToken,
-    ) -> Result<(), crate::agent::error::AgentError> {
-        if self
-            .tool_orchestrator
-            .parameter_reformat_provider
-            .is_empty()
-        {
-            return Ok(());
-        }
-        let budget_secs = self.tool_orchestrator.max_retry_duration_secs;
-        for idx in 0..tool_results.len() {
-            if cancel.is_cancelled() {
-                self.tool_executor.set_skill_env(None);
-                tracing::info!("parameter reformat phase cancelled by user");
-                self.update_metrics(|m| m.cancellations += 1);
-                self.channel.send("[Cancelled]").await?;
-                self.persist_cancelled_tool_results(tool_calls).await;
-                return Ok(());
-            }
-            let needs_reformat = matches!(
-                tool_results[idx],
-                Err(ref e) if e.category().needs_parameter_reformat()
-            );
-            if !needs_reformat {
-                continue;
-            }
-            let tc = &tool_calls[idx];
-            tracing::warn!(
-                tool = %tc.name,
-                "parameter error detected; parameter reformat path is reserved for future \
-                 LLM-based reformat implementation"
-            );
-            // Budget check: a newly created instant always has ~0 elapsed, so this guard
-            // is effectively a no-op today. Kept for structural parity with the planned
-            // LLM-based reformat implementation that will run actual work here.
-            let reformat_start = std::time::Instant::now();
-            if budget_secs > 0 && reformat_start.elapsed().as_secs() >= budget_secs {
-                tracing::warn!(tool = %tc.name, "parameter reformat budget exhausted, skipping");
-                continue;
-            }
-            let _ = self
-                .channel
-                .send_status(&format!(
-                    "Reformat for {} pending provider integration…",
-                    tc.name
-                ))
-                .await;
-            let _ = self.channel.send_status("").await;
-        }
-        Ok(())
-    }
-
-    /// Run all registered pre-execution verifiers against each tool call.
-    ///
-    /// Returns a `Vec<bool>` where `true` at index `i` means call `i` was blocked.
-    /// Blocking fires an audit log entry via the supervisor (fire-and-forget).
-    fn run_pre_execution_verifiers(&mut self, calls: &[ToolCall]) -> Vec<bool> {
-        let mut pre_exec_blocked = vec![false; calls.len()];
-        if self.tool_orchestrator.pre_execution_verifiers.is_empty() {
-            return pre_exec_blocked;
-        }
-        for (idx, call) in calls.iter().enumerate() {
-            let args_value = serde_json::Value::Object(call.params.clone());
-            for verifier in &self.tool_orchestrator.pre_execution_verifiers {
-                match verifier.verify(call.tool_id.as_str(), &args_value) {
-                    zeph_tools::VerificationResult::Allow => {}
-                    zeph_tools::VerificationResult::Block { reason } => {
-                        tracing::warn!(
-                            tool = %call.tool_id,
-                            verifier = verifier.name(),
-                            %reason,
-                            "pre-execution verifier blocked tool call"
-                        );
-                        self.update_metrics(|m| m.pre_execution_blocks += 1);
-                        self.push_security_event(
-                            zeph_common::SecurityEventCategory::PreExecutionBlock,
-                            call.tool_id.as_str(),
-                            format!("{}: {}", verifier.name(), reason),
-                        );
-                        if let Some(ref logger) = self.tool_orchestrator.audit_logger {
-                            let args_json = serde_json::to_string(&args_value).unwrap_or_default();
-                            let entry = zeph_tools::AuditEntry {
-                                timestamp: zeph_tools::chrono_now(),
-                                tool: call.tool_id.clone(),
-                                command: args_json,
-                                result: zeph_tools::AuditResult::Blocked {
-                                    reason: format!("{}: {}", verifier.name(), reason),
-                                },
-                                duration_ms: 0,
-                                error_category: Some("pre_execution_block".to_owned()),
-                                error_domain: Some("security".to_owned()),
-                                error_phase: Some(
-                                    zeph_tools::error_taxonomy::ToolInvocationPhase::Setup
-                                        .label()
-                                        .to_owned(),
-                                ),
-                                claim_source: None,
-                                mcp_server_id: None,
-                                injection_flagged: false,
-                                embedding_anomalous: false,
-                                cross_boundary_mcp_to_acp: false,
-                                adversarial_policy_decision: None,
-                                exit_code: None,
-                                truncated: false,
-                                caller_id: call.caller_id.clone(),
-                                policy_match: None,
-                                correlation_id: None,
-                                vigil_risk: None,
-                                execution_env: None,
-                                resolved_cwd: None,
-                                scope_at_definition: None,
-                                scope_at_dispatch: None,
-                            };
-                            let logger = std::sync::Arc::clone(logger);
-                            self.runtime.lifecycle.supervisor.spawn(
-                                crate::agent::agent_supervisor::TaskClass::Telemetry,
-                                "audit-log",
-                                async move { logger.log(&entry).await },
-                            );
-                        }
-                        pre_exec_blocked[idx] = true;
-                        break;
-                    }
-                    zeph_tools::VerificationResult::Warn { message } => {
-                        tracing::warn!(
-                            tool = %call.tool_id,
-                            verifier = verifier.name(),
-                            %message,
-                            "pre-execution verifier warning (not blocked)"
-                        );
-                        self.update_metrics(|m| m.pre_execution_warnings += 1);
-                        self.push_security_event(
-                            zeph_common::SecurityEventCategory::PreExecutionWarn,
-                            call.tool_id.as_str(),
-                            format!("{}: {}", verifier.name(), message),
-                        );
-                    }
-                }
-            }
-        }
-        pre_exec_blocked
-    }
-
-    /// Score each tool call with the utility gate and return a recommended action per call.
-    ///
-    /// Pre-exec-blocked calls are returned as `ToolCall` to avoid double-counting. Exempt tools
-    /// bypass scoring entirely. For all others, the scorer recommends an action based on context
-    /// (token budget, call history, explicit user request).
-    fn compute_utility_actions(
-        &mut self,
-        calls: &[ToolCall],
-        pre_exec_blocked: &[bool],
-    ) -> Vec<zeph_tools::UtilityAction> {
-        #[allow(clippy::cast_possible_truncation)]
-        let tokens_consumed =
-            usize::try_from(self.runtime.providers.cached_prompt_tokens).unwrap_or(usize::MAX);
-        // token_budget = 0 signals "unknown" to UtilityContext — cost component is zeroed.
-        let token_budget: usize = 0;
-        let tool_calls_this_turn = self.tool_orchestrator.recent_tool_calls.len();
-        // Detect explicit tool request from the last user message text only.
-        // We only read MessagePart::Text parts so tool outputs/thinking blocks are excluded.
-        let explicit_request = self
-            .msg
-            .messages
-            .iter()
-            .rfind(|m| m.role == zeph_llm::provider::Role::User)
-            .is_some_and(|m| {
-                let text = if m.parts.is_empty() {
-                    m.content.clone()
-                } else {
-                    m.parts
-                        .iter()
-                        .filter_map(|p| {
-                            if let zeph_llm::provider::MessagePart::Text { text } = p {
-                                Some(text.as_str())
-                            } else {
-                                None
-                            }
-                        })
-                        .collect::<Vec<_>>()
-                        .join(" ")
-                };
-                zeph_tools::has_explicit_tool_request(&text)
-            });
-        calls
-            .iter()
-            .enumerate()
-            .map(|(idx, call)| {
-                if pre_exec_blocked[idx] {
-                    return zeph_tools::UtilityAction::ToolCall;
-                }
-                if self
-                    .tool_orchestrator
-                    .utility_scorer
-                    .is_exempt(call.tool_id.as_str())
-                {
-                    return zeph_tools::UtilityAction::ToolCall;
-                }
-                let ctx = zeph_tools::UtilityContext {
-                    tool_calls_this_turn: tool_calls_this_turn + idx,
-                    tokens_consumed,
-                    token_budget,
-                    user_requested: explicit_request,
-                };
-                let score = self.tool_orchestrator.utility_scorer.score(call, &ctx);
-                let action = self
-                    .tool_orchestrator
-                    .utility_scorer
-                    .recommend_action(score.as_ref(), &ctx);
-                tracing::debug!(
-                    tool = %call.tool_id,
-                    score = ?score.as_ref().map(|s| s.total),
-                    threshold = self.tool_orchestrator.utility_scorer.threshold(),
-                    action = ?action,
-                    "utility gate: action recommended"
-                );
-                if action != zeph_tools::UtilityAction::ToolCall {
-                    tracing::info!(
-                        tool = %call.tool_id,
-                        action = ?action,
-                        "utility gate: non-execute action"
-                    );
-                }
-                // Record call regardless so subsequent calls in this batch see it as prior.
-                self.tool_orchestrator.utility_scorer.record_call(call);
-                action
-            })
-            .collect()
-    }
-
-    #[cfg_attr(
-        feature = "profiling",
-        tracing::instrument(name = "agent.tool_loop", skip_all)
-    )]
-    pub(super) async fn handle_native_tool_calls(
-        &mut self,
-        text: Option<&str>,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-    ) -> Result<(), crate::agent::error::AgentError> {
-        let t_tool_exec = std::time::Instant::now();
-        tracing::debug!("turn timing: tool_exec start");
-        // Scan for image-exfiltration in accompanying text, send to channel, persist
-        // the assistant ToolUse message.
-        self.push_assistant_tool_use_message(text, tool_calls)
-            .await?;
-
-        // Build calls, assign IDs, run exfiltration guard, gate checks (pre-exec/utility/
-        // quota/repeat/cache), and inject skill env. Extracted to keep this function under
-        // the clippy line limit.
-        let ToolDispatchContext {
-            calls,
-            tool_call_ids,
-            mut tool_started_ats,
-            pre_exec_blocked,
-            utility_actions,
-            quota_blocked,
-            args_hashes,
-            repeat_blocked,
-            cache_hits,
-        } = self.prepare_tool_dispatch(tool_calls);
-
-        let max_retries = self.tool_orchestrator.max_tool_retries;
-        // Clamp to 1 to prevent Semaphore(0) deadlock when config is set to 0.
-        let max_parallel = self.runtime.config.timeouts.max_parallel_tools.max(1);
-        let cancel = self.runtime.lifecycle.cancel_token.clone();
-
-        // Causal IPI pre-probe: record behavioral baseline before tool batch dispatch.
-        let causal_pre_response = self.run_causal_pre_probe().await;
-
-        // Phase 1: Tiered parallel execution bounded by a shared semaphore.
-        // Extracted to run_tier_execution_loop to satisfy the line-count limit.
-        // Returns None when the user cancelled (caller must return Ok(())).
-        let tier_data = self
-            .run_tier_execution_loop(
-                tool_calls,
-                &calls,
-                &pre_exec_blocked,
-                &utility_actions,
-                quota_blocked,
-                &args_hashes,
-                &repeat_blocked,
-                &cache_hits,
-                max_parallel,
-                &cancel,
-                &tool_call_ids,
-                &mut tool_started_ats,
-            )
-            .await?;
-
-        // Unpack tier execution output. None means the user cancelled — return early.
-        let Some(TierLoopData {
-            mut tool_results,
-            pending_focus_checkpoint,
-            pending_system_hints,
-        }) = tier_data
-        else {
-            return Ok(());
-        };
-
-        // Phases 2a / 2 / 3: confirmation, transient retry, parameter reformat.
-        // Each phase may return early on cancellation (Ok(())) or propagate channel errors.
-        self.run_post_dispatch_phases(tool_calls, &calls, &mut tool_results, max_retries, &cancel)
-            .await?;
-
-        // Process results, persist messages, run LSP hooks, fire deferred reflection.
-        // Also clears skill env and syncs cache counters after execution.
-        // Extracted to process_tool_result_batch to satisfy the line-count limit.
-        self.process_tool_result_batch(
-            tool_calls,
-            &tool_call_ids,
-            &tool_started_ats,
-            tool_results,
-            causal_pre_response,
-            pending_focus_checkpoint,
-            pending_system_hints,
-        )
-        .await?;
-
-        let tool_exec_ms = u64::try_from(t_tool_exec.elapsed().as_millis()).unwrap_or(u64::MAX);
-        tracing::debug!(ms = tool_exec_ms, "turn timing: tool_exec done");
-        self.runtime.metrics.pending_timings.tool_exec_ms = self
-            .runtime
-            .metrics
-            .pending_timings
-            .tool_exec_ms
-            .saturating_add(tool_exec_ms);
-
-        Ok(())
-    }
-
-    /// Scan the optional text accompanying a `ToolUse` LLM response for injection patterns,
-    /// send it to the channel, then persist the assistant `ToolUse` message to history.
-    ///
-    /// Separated from `handle_native_tool_calls` to keep that function under the line limit.
-    async fn push_assistant_tool_use_message(
-        &mut self,
-        text: Option<&str>,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-    ) -> Result<(), crate::agent::error::AgentError> {
-        // S4: scan text accompanying ToolUse responses for markdown image exfiltration.
-        let cleaned_text: Option<String> = if let Some(t) = text
-            && !t.is_empty()
-        {
-            Some(self.scan_output_and_warn(t))
-        } else {
-            None
-        };
-
-        if let Some(ref t) = cleaned_text
-            && !t.is_empty()
-        {
-            let display = self.maybe_redact(t);
-            self.channel.send(&display).await?;
-        }
-
-        let mut parts: Vec<MessagePart> = Vec::new();
-        if let Some(ref t) = cleaned_text
-            && !t.is_empty()
-        {
-            parts.push(MessagePart::Text { text: t.clone() });
-        }
-        for tc in tool_calls {
-            parts.push(MessagePart::ToolUse {
-                id: tc.id.clone(),
-                name: tc.name.to_string(),
-                input: tc.input.clone(),
-            });
-        }
-        let assistant_msg = Message::from_parts(Role::Assistant, parts);
-        self.persist_message(
-            Role::Assistant,
-            &assistant_msg.content,
-            &assistant_msg.parts,
-            false,
-        )
-        .await;
-        self.push_message(assistant_msg);
-        if let (Some(id), Some(last)) = (
-            self.msg.last_persisted_message_id,
-            self.msg.messages.last_mut(),
-        ) {
-            last.metadata.db_id = Some(id);
-        }
-        Ok(())
-    }
-
-    /// Build the canonical call list from `tool_calls`, assign stable IDs, run the exfiltration
-    /// guard, and compute all per-call gate results (pre-exec block, utility action, quota,
-    /// repeat detection, cache lookup). Injects active skill secrets before returning.
-    ///
-    /// Separated from `handle_native_tool_calls` to keep that function under the line limit.
-    fn prepare_tool_dispatch(
-        &mut self,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-    ) -> ToolDispatchContext {
-        let tafc_enabled = self.tool_orchestrator.tafc.enabled;
-        // When the orchestration scheduler has set a named execution environment for the
-        // current task, inject it into every ToolCall so ShellExecutor::resolve_context
-        // uses the right env/cwd without the LLM having to supply it.
-        let task_ctx: Option<ExecutionContext> = self
-            .services
-            .orchestration
-            .task_execution_env
-            .as_deref()
-            .map(|name| ExecutionContext::default().with_name(name));
-        let calls: Vec<ToolCall> = tool_calls
-            .iter()
-            .filter_map(|tc| {
-                let mut params: serde_json::Map<String, serde_json::Value> =
-                    if let serde_json::Value::Object(map) = &tc.input {
-                        map.clone()
-                    } else {
-                        serde_json::Map::new()
-                    };
-                if tafc_enabled && strip_tafc_fields(&mut params, tc.name.as_str()).is_err() {
-                    // Model produced only think fields — skip this tool call.
-                    return None;
-                }
-                Some(ToolCall {
-                    tool_id: tc.name.clone(),
-                    params,
-                    caller_id: None,
-                    context: task_ctx.clone(),
-                })
-            })
-            .collect();
-
-        // Assign stable IDs before execution so ToolStart and ToolOutput share the same ID.
-        let tool_call_ids: Vec<String> = tool_calls
-            .iter()
-            .map(|_| uuid::Uuid::new_v4().to_string())
-            .collect();
-        // tool_started_ats is populated per-tier just before each tier's join_all so that
-        // audit timestamps reflect actual execution start rather than pre-build time.
-        let tool_started_ats: Vec<std::time::Instant> =
-            vec![std::time::Instant::now(); tool_calls.len()];
-
-        self.check_exfiltration_urls(tool_calls);
-
-        // Pre-execution verification (TrustBench pattern, issue #1630).
-        // Runs after exfiltration guard (flag-only) and before repeat-detection.
-        // Block: return synthetic error result for this call without executing.
-        // Warn: log + emit security event + continue execution.
-        let pre_exec_blocked = self.run_pre_execution_verifiers(&calls);
-
-        // Utility gate: score each call and recommend an action (#2477).
-        // user_requested is detected from the last user message only — never from LLM content or
-        // tool outputs to prevent prompt-injection bypass (C2 fix).
-        let utility_actions = self.compute_utility_actions(&calls, &pre_exec_blocked);
-
-        // Per-session quota check. Counted once per logical dispatch batch (M3: retries do not
-        // consume additional quota slots). When exceeded, all calls in this batch are quota-blocked.
-        let quota_blocked = if let Some(max) = self.tool_orchestrator.check_quota() {
-            tracing::warn!(
-                max,
-                count = self.tool_orchestrator.session_tool_call_count,
-                "tool call quota exceeded for session"
-            );
-            true
-        } else {
-            // Increment before the retry loop — each call in this batch counts as one logical call.
-            self.tool_orchestrator.session_tool_call_count = self
-                .tool_orchestrator
-                .session_tool_call_count
-                .saturating_add(u32::try_from(calls.len()).unwrap_or(u32::MAX));
-            false
-        };
-
-        // Build args hashes and check for repeats. Blocked calls get a pre-built error result.
-        let args_hashes: Vec<u64> = calls.iter().map(|c| tool_args_hash(&c.params)).collect();
-        let repeat_blocked: Vec<bool> = calls
-            .iter()
-            .zip(args_hashes.iter())
-            .map(|(call, &hash)| {
-                let blocked = self
-                    .tool_orchestrator
-                    .is_repeat(call.tool_id.as_str(), hash);
-                if blocked {
-                    tracing::warn!(
-                        tool = %call.tool_id,
-                        "[repeat-detect] identical tool call detected, skipping execution"
-                    );
-                }
-                blocked
-            })
-            .collect();
-        // Repeat-detection (CRIT-3): push LLM-initiated calls BEFORE execution.
-        // Cache hits are also pushed here (P1 invariant): a cached tool called N times must
-        // still trigger repeat-detection to prevent infinite loops if the LLM keeps requesting it.
-        for (call, &hash) in calls.iter().zip(args_hashes.iter()) {
-            self.tool_orchestrator
-                .push_tool_call(call.tool_id.as_str(), hash);
-        }
-
-        // Cache lookup: for each non-repeat, cacheable call, check result cache before dispatch.
-        // Hits are stored as pre-built results; cache store happens after join_all completes.
-        let cache_hits: Vec<Option<zeph_tools::ToolOutput>> = calls
-            .iter()
-            .zip(args_hashes.iter())
-            .zip(repeat_blocked.iter())
-            .map(|((call, &hash), &blocked)| {
-                if blocked || !zeph_tools::is_cacheable(call.tool_id.as_str()) {
-                    return None;
-                }
-                let key = zeph_tools::CacheKey::new(call.tool_id.as_str(), hash);
-                self.tool_orchestrator.result_cache.get(&key)
-            })
-            .collect();
-
-        // Inject active skill secrets before tool execution.
-        self.inject_active_skill_env();
-
-        ToolDispatchContext {
-            calls,
-            tool_call_ids,
-            tool_started_ats,
-            pre_exec_blocked,
-            utility_actions,
-            quota_blocked,
-            args_hashes,
-            repeat_blocked,
-            cache_hits,
-        }
-    }
-
-    /// Validate tool call arguments against URLs seen in flagged untrusted content.
-    ///
-    /// This is a flag-only check — suspicious URLs are logged and metered but do not block
-    /// execution. Extracted from `prepare_tool_dispatch` to keep that function under the
-    /// clippy line limit.
-    fn check_exfiltration_urls(&mut self, tool_calls: &[zeph_llm::provider::ToolUseRequest]) {
-        for tc in tool_calls {
-            let args_json = tc.input.to_string();
-            let url_events = self
-                .services
-                .security
-                .exfiltration_guard
-                .validate_tool_call(
-                    tc.name.as_str(),
-                    &args_json,
-                    &self.services.security.flagged_urls,
-                );
-            if !url_events.is_empty() {
-                tracing::warn!(
-                    tool = %tc.name,
-                    count = url_events.len(),
-                    "exfiltration guard: suspicious URLs in tool arguments (flag-only, not blocked)"
-                );
-                self.update_metrics(|m| {
-                    m.exfiltration_tool_urls_flagged += url_events.len() as u64;
-                });
-                self.push_security_event(
-                    zeph_common::SecurityEventCategory::ExfiltrationBlock,
-                    tc.name.as_str(),
-                    format!(
-                        "{} suspicious URL(s) flagged in tool args",
-                        url_events.len()
-                    ),
-                );
-            }
-        }
-    }
-
-    /// Run the causal IPI pre-probe to record the behavioral baseline before tool dispatch.
-    ///
-    /// Returns the probe response paired with the context summary so the post-probe can
-    /// compare the two. Returns `None` when the analyzer is not configured or the probe fails.
-    async fn run_causal_pre_probe(&mut self) -> Option<(String, String)> {
-        let analyzer = self.services.security.causal_analyzer.as_ref()?;
-        let context_summary = self.build_causal_context_summary();
-        match analyzer.probe(&context_summary).await {
-            Ok(resp) => Some((resp, context_summary)),
-            Err(e) => {
-                tracing::warn!(error = %e, "causal IPI pre-probe failed, skipping analysis");
-                None
-            }
-        }
-    }
-
-    /// Execute tool calls in topological tiers (Phase 1).
-    ///
-    /// Builds a dependency DAG, partitions into tiers, and runs each tier with bounded
-    /// concurrency via a shared semaphore. MCP elicitation events are drained during the
-    /// join to prevent deadlocks.
-    ///
-    /// Returns `None` when the user cancelled (the caller must return `Ok(())`), or
-    /// `Some(TierLoopData)` with the collected results on success.
-    #[allow(clippy::too_many_arguments, clippy::too_many_lines)]
-    async fn run_tier_execution_loop(
-        &mut self,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-        calls: &[ToolCall],
-        pre_exec_blocked: &[bool],
-        utility_actions: &[zeph_tools::UtilityAction],
-        quota_blocked: bool,
-        args_hashes: &[u64],
-        repeat_blocked: &[bool],
-        cache_hits: &[Option<zeph_tools::ToolOutput>],
-        max_parallel: usize,
-        cancel: &tokio_util::sync::CancellationToken,
-        tool_call_ids: &[String],
-        tool_started_ats: &mut [std::time::Instant],
-    ) -> Result<TierLoopOutput, crate::agent::error::AgentError> {
-        // Build a dependency DAG over tool_use_id references in call arguments. When the
-        // DAG is trivial (no dependencies — the common case), we execute all calls in a
-        // single tier with zero overhead. When dependencies exist, we partition calls into
-        // topological tiers and execute each tier in parallel, awaiting the previous tier
-        // before starting the next.
-        //
-        // ToolStartEvent is sent at the beginning of each tier so the UI reflects actual
-        // execution start time rather than pre-build time.
-        let semaphore = std::sync::Arc::new(tokio::sync::Semaphore::new(max_parallel));
-        let dag = super::tool_call_dag::ToolCallDag::build(tool_calls);
-        let trivial = dag.is_trivial();
-        let tiers = dag.tiers();
-        let tier_count = tiers.len();
-        // Clone the Arc before the mutable borrow loop so try_commit can be called without
-        // holding a borrow on self across await points.
-        let speculation_engine = self.services.speculation_engine.clone();
-        tracing::debug!(
-            trivial,
-            tier_count,
-            tool_count = tool_calls.len(),
-            "tool dispatch: partitioned into tiers"
-        );
-
-        // Pre-allocate result vector; slots are filled as tiers complete.
-        let mut tool_results: Vec<Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>> =
-            (0..tool_calls.len()).map(|_| Ok(None)).collect();
-
-        // Pre-process focus tool calls (#1850) and compress_context (#2218).
-        // These need &mut self and cannot run inside the parallel tier futures.
-        // Pre-populate their results so the tier loop skips them.
-        let pending_focus_checkpoint = self
-            .preprocess_focus_compress_calls(tool_calls, &mut tool_results)
-            .await;
-
-        // Track which indices have a failed/ConfirmationRequired prerequisite so that
-        // dependent calls in later tiers receive a synthetic error instead of executing.
-        // IMP-02: ConfirmationRequired is treated as a failure for dependency propagation —
-        // a dependent tool must not proceed when its prerequisite is awaiting user approval.
-        let mut failed_ids: std::collections::HashSet<String> = std::collections::HashSet::new();
-        // Utility gate hints (Retrieve/Verify) are deferred so they are pushed after
-        // User(tool_results), maintaining valid OpenAI message ordering (#2615).
-        let mut pending_system_hints: Vec<String> = Vec::new();
-
-        for (tier_idx, tier) in tiers.into_iter().enumerate() {
-            if cancel.is_cancelled() {
-                self.tool_executor.set_skill_env(None);
-                tracing::info!("tool execution cancelled by user");
-                self.update_metrics(|m| m.cancellations += 1);
-                self.channel.send("[Cancelled]").await?;
-                self.persist_cancelled_tool_results(tool_calls).await;
-                return Ok(None);
-            }
-
-            if tier_count > 1 {
-                let _ = self
-                    .channel
-                    .send_status(&format!(
-                        "Executing tools (tier {}/{})\u{2026}",
-                        tier_idx + 1,
-                        tier_count
-                    ))
-                    .await;
-            }
-
-            // Pre-scan: commit speculative handles and emit speculative ToolStartEvents.
-            let speculative_commits = self
-                .commit_speculative_tier(
-                    &tier.indices,
-                    calls,
-                    tool_calls,
-                    tool_call_ids,
-                    tool_started_ats,
-                    speculation_engine.as_ref(),
-                )
-                .await?;
-
-            // Stamp execution start time and send ToolStartEvent for non-committed calls (§3.7).
-            let non_committed_indices: Vec<usize> = tier
-                .indices
-                .iter()
-                .copied()
-                .filter(|idx| !speculative_commits.contains_key(idx))
-                .collect();
-            self.stamp_and_send_tier_start(
-                &non_committed_indices,
-                tool_calls,
-                tool_call_ids,
-                tool_started_ats,
-            )
-            .await?;
-
-            // Build futures for non-committed calls in this tier.
-            let mut tier_futs = self
-                .build_tier_call_futures(
-                    tool_calls,
-                    calls,
-                    &non_committed_indices,
-                    &dag,
-                    &failed_ids,
-                    quota_blocked,
-                    pre_exec_blocked,
-                    utility_actions,
-                    repeat_blocked,
-                    cache_hits,
-                    &semaphore,
-                    &mut pending_system_hints,
-                )
-                .await?;
-
-            // Inject committed speculative results as ready futures.
-            for (idx, result) in speculative_commits {
-                tier_futs.push((idx, Box::pin(std::future::ready(result))));
-            }
-
-            // Execute futures concurrently with cancellation and MCP elicitation drain.
-            let (indices, futs): (Vec<usize>, Vec<ToolExecFut>) = tier_futs.into_iter().unzip();
-            let Some(tier_results) = self.execute_tier_join(futs, cancel, tool_calls).await? else {
-                return Ok(None);
-            };
-
-            // Store results, update dependency graph, and run after_tool hooks.
-            self.apply_tier_results(
-                indices,
-                tier_results,
-                tool_calls,
-                calls,
-                cache_hits,
-                args_hashes,
-                &mut failed_ids,
-                &mut tool_results,
-            )
-            .await;
-
-            if tier_count > 1 {
-                let _ = self.channel.send_status("").await;
-            }
-        }
-
-        // Pad with empty results if needed (defensive; should not happen).
-        while tool_results.len() < tool_calls.len() {
-            tool_results.push(Ok(None));
-        }
-
-        Ok(Some(TierLoopData {
-            tool_results,
-            pending_focus_checkpoint,
-            pending_system_hints,
-        }))
-    }
-
-    /// Pre-process focus and `compress_context` tool calls before the tier loop.
-    ///
-    /// These tools require `&mut self` and cannot run inside spawned tier futures.
-    /// Results are pre-populated in `tool_results` so the tier loop skips them.
-    /// Returns the pending focus checkpoint message, if one was produced.
-    async fn preprocess_focus_compress_calls(
-        &mut self,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-        tool_results: &mut [Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>],
-    ) -> Option<zeph_llm::provider::Message> {
-        let mut pending_focus_checkpoint: Option<zeph_llm::provider::Message> = None;
-        for (idx, tc) in tool_calls.iter().enumerate() {
-            let is_focus_tool = self.services.focus.config.enabled
-                && (tc.name == "start_focus" || tc.name == "complete_focus");
-            let is_compress = tc.name == "compress_context";
-            if is_focus_tool || is_compress {
-                let result = if is_compress {
-                    self.handle_compress_context().await
-                } else {
-                    let (text, maybe_checkpoint) =
-                        self.handle_focus_tool(tc.name.as_str(), &tc.input);
-                    if let Some(cp) = maybe_checkpoint {
-                        pending_focus_checkpoint = Some(cp);
-                    }
-                    text
-                };
-                tool_results[idx] = Ok(Some(skipped_output(tc.name.clone(), result)));
-            }
-        }
-        pending_focus_checkpoint
-    }
-
-    /// Mark the execution start time for each call in the tier and send `ToolStartEvent`s.
-    ///
-    /// Called once per tier before futures are dispatched so that TUI timing is accurate.
-    async fn stamp_and_send_tier_start(
-        &mut self,
-        tier_indices: &[usize],
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-        tool_call_ids: &[String],
-        tool_started_ats: &mut [std::time::Instant],
-    ) -> Result<(), crate::agent::error::AgentError> {
-        let tier_start = std::time::Instant::now();
-        for &idx in tier_indices {
-            tool_started_ats[idx] = tier_start;
-        }
-        for &idx in tier_indices {
-            let tc = &tool_calls[idx];
-            self.channel
-                .send_tool_start(ToolStartEvent {
-                    tool_name: tc.name.clone(),
-                    tool_call_id: tool_call_ids[idx].clone(),
-                    params: Some(tc.input.clone()),
-                    parent_tool_use_id: self.services.session.parent_tool_use_id.clone(),
-                    started_at: std::time::Instant::now(),
-                    speculative: false,
-                    sandbox_profile: None,
-                })
-                .await?;
-        }
-        Ok(())
-    }
-
-    /// Attempt to commit speculative handles for all calls in a tier, stamp their start times,
-    /// and emit `ToolStartEvent { speculative: true }` for each committed call.
-    ///
-    /// Returns a map from call index to the committed result. Indices present in the map must be
-    /// excluded from `stamp_and_send_tier_start` and `build_tier_call_futures` — their results are
-    /// injected directly as ready futures instead.
-    async fn commit_speculative_tier(
-        &mut self,
-        tier_indices: &[usize],
-        calls: &[ToolCall],
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-        tool_call_ids: &[String],
-        tool_started_ats: &mut [std::time::Instant],
-        engine: Option<&std::sync::Arc<crate::agent::speculative::SpeculationEngine>>,
-    ) -> Result<
-        std::collections::HashMap<
-            usize,
-            Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>,
-        >,
-        crate::agent::error::AgentError,
-    > {
-        let mut commits: std::collections::HashMap<
-            usize,
-            Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>,
-        > = std::collections::HashMap::new();
-
-        let Some(engine) = engine else {
-            return Ok(commits);
-        };
-
-        for &idx in tier_indices {
-            let Some(result) = engine.try_commit(&calls[idx]).await else {
-                continue;
-            };
-            if let Err(ref e) = result {
-                tracing::warn!(
-                    tool = %calls[idx].tool_id,
-                    error = %e,
-                    "speculative commit returned Err — result will be used as-is"
-                );
-                // Invariant: ConfirmationRequired must never reach the commit boundary —
-                // try_dispatch guards against it at dispatch time via requires_confirmation_erased.
-                #[cfg(debug_assertions)]
-                if matches!(e, zeph_tools::ToolError::ConfirmationRequired { .. }) {
-                    tracing::error!(
-                        tool = %calls[idx].tool_id,
-                        "invariant violated: committed speculative result is ConfirmationRequired"
-                    );
-                }
-            }
-            // M1: stamp actual dispatch time so build_tool_output_messages computes correct elapsed.
-            tool_started_ats[idx] = std::time::Instant::now();
-            commits.insert(idx, result);
-        }
-
-        // Emit ToolStartEvent with speculative: true for all committed calls.
-        for &idx in tier_indices {
-            if commits.contains_key(&idx) {
-                let tc = &tool_calls[idx];
-                self.channel
-                    .send_tool_start(ToolStartEvent {
-                        tool_name: tc.name.clone(),
-                        tool_call_id: tool_call_ids[idx].clone(),
-                        params: Some(tc.input.clone()),
-                        parent_tool_use_id: self.services.session.parent_tool_use_id.clone(),
-                        started_at: tool_started_ats[idx],
-                        speculative: true,
-                        sandbox_profile: None,
-                    })
-                    .await?;
-            }
-        }
-
-        Ok(commits)
-    }
-
-    /// Apply the utility gate decision for a single tool call.
-    ///
-    /// Returns `Some((idx, fut))` when the gate fires and the call should be short-circuited
-    /// (the caller must push the future and `continue`), or `None` to proceed normally.
-    /// Mutates `pending_system_hints` for Retrieve/Verify actions.
-    async fn handle_utility_gate(
-        &mut self,
-        idx: usize,
-        tc: &zeph_llm::provider::ToolUseRequest,
-        utility_actions: &[zeph_tools::UtilityAction],
-        pending_system_hints: &mut Vec<String>,
-    ) -> Result<Option<(usize, ToolExecFut)>, crate::agent::error::AgentError> {
-        match utility_actions[idx] {
-            zeph_tools::UtilityAction::ToolCall => Ok(None),
-            zeph_tools::UtilityAction::Respond => {
-                let _ = self
-                    .channel
-                    .send_status(&format!("Utility action: Respond ({})", tc.name))
-                    .await;
-                Ok(Some(ready_fut(
-                    idx,
-                    skipped_output(
-                        tc.name.clone(),
-                        format!(
-                            "[skipped] Tool call to {} skipped — utility policy recommends a \
-                             direct response without further tool use.",
-                            tc.name
-                        ),
-                    ),
-                )))
-            }
-            zeph_tools::UtilityAction::Retrieve => {
-                let _ = self
-                    .channel
-                    .send_status(&format!("Utility action: Retrieve ({})", tc.name))
-                    .await;
-                // Inject a system message directing the LLM to retrieve context first (#2620).
-                pending_system_hints.push(format!(
-                    "[utility:retrieve] Before executing the '{}' tool, retrieve \
-                     relevant context via memory_search or a related lookup to ensure \
-                     the call is well-targeted. After retrieving context, you MUST call \
-                     the '{}' tool again with the same arguments.",
-                    tc.name, tc.name
-                ));
-                Ok(Some(ready_fut(
-                    idx,
-                    skipped_output(
-                        tc.name.clone(),
-                        format!(
-                            "[skipped] Tool call to {} skipped — utility policy recommends \
-                             retrieving additional context first.",
-                            tc.name
-                        ),
-                    ),
-                )))
-            }
-            zeph_tools::UtilityAction::Verify => {
-                let _ = self
-                    .channel
-                    .send_status(&format!("Utility action: Verify ({})", tc.name))
-                    .await;
-                pending_system_hints.push(format!(
-                    "[utility:verify] Before executing the '{}' tool again, verify \
-                     the result of the previous tool call to confirm it is correct \
-                     and that further tool use is necessary.",
-                    tc.name
-                ));
-                Ok(Some(ready_fut(
-                    idx,
-                    skipped_output(
-                        tc.name.clone(),
-                        format!(
-                            "[skipped] Tool call to {} skipped — utility policy recommends \
-                             verifying the previous result first.",
-                            tc.name
-                        ),
-                    ),
-                )))
-            }
-            zeph_tools::UtilityAction::Stop => {
-                let _ = self
-                    .channel
-                    .send_status(&format!("Utility action: Stop ({})", tc.name))
-                    .await;
-                let threshold = self.tool_orchestrator.utility_scorer.threshold();
-                Ok(Some(ready_fut(
-                    idx,
-                    skipped_output(
-                        tc.name.clone(),
-                        format!(
-                            "[stopped] Tool call to {} halted by the utility gate — \
-                             budget exhausted or score below threshold {threshold:.2}.",
-                            tc.name
-                        ),
-                    ),
-                )))
-            }
-        }
-    }
-
-    /// Run `RuntimeLayer::before_tool` hooks for a single call.
-    ///
-    /// Returns `Some((idx, fut))` when a hook short-circuits execution (the caller must push and
-    /// `continue`), or `None` when all hooks pass and execution should proceed normally.
-    /// `PermissionDenied` hooks are fired asynchronously when a layer blocks the call.
-    async fn run_before_tool_hooks(
-        &mut self,
-        idx: usize,
-        tc: &zeph_llm::provider::ToolUseRequest,
-        call: &ToolCall,
-    ) -> Option<(usize, ToolExecFut)> {
-        if self.runtime.config.layers.is_empty() {
-            return None;
-        }
-        let conv_id_str = self
-            .services
-            .memory
-            .persistence
-            .conversation_id
-            .map(|id| id.0.to_string());
-        let ctx = crate::runtime_layer::LayerContext {
-            conversation_id: conv_id_str.as_deref(),
-            turn_number: u32::try_from(self.services.sidequest.turn_counter).unwrap_or(u32::MAX),
-        };
-        let mut sc_result: crate::runtime_layer::BeforeToolResult = None;
-        for layer in &self.runtime.config.layers {
-            let hook_result = std::panic::AssertUnwindSafe(layer.before_tool(&ctx, call))
-                .catch_unwind()
-                .await;
-            match hook_result {
-                Ok(Some(r)) => {
-                    sc_result = Some(r);
-                    break;
-                }
-                Ok(None) => {}
-                Err(_) => {
-                    tracing::warn!("RuntimeLayer::before_tool panicked, continuing");
-                }
-            }
-        }
-        let r = sc_result?;
-        // Fire PermissionDenied hooks (fail_open: hook errors are logged, not fatal).
-        let pd_hooks = self.services.session.hooks_config.permission_denied.clone();
-        if !pd_hooks.is_empty() {
-            let _span =
-                tracing::info_span!("core.hooks.permission_denied", tool = %tc.name).entered();
-            let mut env = std::collections::HashMap::new();
-            env.insert("ZEPH_DENIED_TOOL".to_owned(), tc.name.to_string());
-            env.insert("ZEPH_DENY_REASON".to_owned(), r.reason.clone());
-            let dispatch = self.mcp_dispatch();
-            let mcp: Option<&dyn zeph_subagent::McpDispatch> = dispatch
-                .as_ref()
-                .map(|d| d as &dyn zeph_subagent::McpDispatch);
-            // TODO: implement retry-on-{"retry":true} stdout signal (#3292)
-            if let Err(e) = zeph_subagent::hooks::fire_hooks(&pd_hooks, &env, mcp).await {
-                tracing::warn!(error = %e, tool = %tc.name, "PermissionDenied hook failed");
-            }
-        }
-        Some((idx, Box::pin(std::future::ready(r.result))))
-    }
-
-    /// Check the static per-call gates in order: dependency failure, quota, pre-exec block,
-    /// utility gate action, and repeat detection.
-    ///
-    /// Returns `Some((idx, fut))` when any gate fires (the caller must push and `continue`),
-    /// or `None` when all gates pass and execution should proceed.
-    #[allow(clippy::too_many_arguments)]
-    async fn check_call_gates(
-        &mut self,
-        idx: usize,
-        tc: &zeph_llm::provider::ToolUseRequest,
-        has_failed_dep: bool,
-        quota_blocked: bool,
-        pre_exec_blocked: &[bool],
-        utility_actions: &[zeph_tools::UtilityAction],
-        repeat_blocked: &[bool],
-        pending_system_hints: &mut Vec<String>,
-    ) -> Result<Option<(usize, ToolExecFut)>, crate::agent::error::AgentError> {
-        if has_failed_dep {
-            return Ok(Some(ready_fut(
-                idx,
-                skipped_output(
-                    tc.name.clone(),
-                    "[error] Skipped: a prerequisite tool failed or requires confirmation",
-                ),
-            )));
-        }
-        if quota_blocked {
-            let max = self
-                .tool_orchestrator
-                .max_tool_calls_per_session
-                .unwrap_or(0);
-            return Ok(Some(ready_fut(
-                idx,
-                skipped_output(
-                    tc.name.clone(),
-                    format!(
-                        "[error] Tool call quota exceeded (session limit: {max} calls). \
-                         No further tool calls are allowed this session."
-                    ),
-                ),
-            )));
-        }
-        if pre_exec_blocked[idx] {
-            return Ok(Some(ready_fut(
-                idx,
-                skipped_output(
-                    tc.name.clone(),
-                    format!(
-                        "[error] Tool call to {} was blocked by pre-execution verifier. \
-                         The requested operation is not permitted.",
-                        tc.name
-                    ),
-                ),
-            )));
-        }
-        if let Some(fut) = self
-            .handle_utility_gate(idx, tc, utility_actions, pending_system_hints)
-            .await?
-        {
-            return Ok(Some(fut));
-        }
-        if repeat_blocked[idx] {
-            return Ok(Some(ready_fut(
-                idx,
-                skipped_output(
-                    tc.name.clone(),
-                    format!(
-                        "[error] Repeated identical call to {} detected. \
-                         Use different arguments or a different approach.",
-                        tc.name
-                    ),
-                ),
-            )));
-        }
-        Ok(None)
-    }
-
-    /// Build the execution futures for one DAG tier, applying all per-call gate checks.
-    ///
-    /// For each call in `tier_indices`, the function checks (in order): DAG dependency
-    /// failure, session quota, pre-exec block, utility gate, repeat detection, result cache,
-    /// rate limiter, and `RuntimeLayer::before_tool` hooks. Any gate that fires returns a
-    /// pre-resolved synthetic future. Calls that pass all gates get a semaphore-guarded
-    /// executor future. Returns `Err` only for channel errors from utility-gate status sends.
-    #[allow(clippy::too_many_arguments, clippy::ptr_arg)]
-    async fn build_tier_call_futures(
-        &mut self,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-        calls: &[ToolCall],
-        tier_indices: &[usize],
-        dag: &super::tool_call_dag::ToolCallDag,
-        failed_ids: &std::collections::HashSet<String>,
-        quota_blocked: bool,
-        pre_exec_blocked: &[bool],
-        utility_actions: &[zeph_tools::UtilityAction],
-        repeat_blocked: &[bool],
-        cache_hits: &[Option<zeph_tools::ToolOutput>],
-        semaphore: &std::sync::Arc<tokio::sync::Semaphore>,
-        pending_system_hints: &mut Vec<String>,
-    ) -> Result<Vec<(usize, ToolExecFut)>, crate::agent::error::AgentError> {
-        let tier_tool_names: Vec<&str> = tier_indices
-            .iter()
-            .map(|&i| tool_calls[i].name.as_str())
-            .collect();
-        let rate_results = self
-            .runtime
-            .config
-            .rate_limiter
-            .check_batch(&tier_tool_names);
-
-        let mut tier_futs: Vec<(usize, ToolExecFut)> = Vec::with_capacity(tier_indices.len());
-        for (tier_local_idx, &idx) in tier_indices.iter().enumerate() {
-            let tc = &tool_calls[idx];
-            let call = &calls[idx];
-
-            // Skip focus tools and compress_context — pre-handled before the tier loop.
-            if tc.name == "compress_context"
-                || (self.services.focus.config.enabled
-                    && (tc.name == "start_focus" || tc.name == "complete_focus"))
-            {
-                continue;
-            }
-
-            // Check static gates: dep failure, quota, pre-exec block, utility gate, repeat.
-            let has_failed_dep = dag
-                .string_values_for(idx)
-                .iter()
-                .any(|v| failed_ids.contains(v));
-            if let Some(fut) = self
-                .check_call_gates(
-                    idx,
-                    tc,
-                    has_failed_dep,
-                    quota_blocked,
-                    pre_exec_blocked,
-                    utility_actions,
-                    repeat_blocked,
-                    pending_system_hints,
-                )
-                .await?
-            {
-                tier_futs.push(fut);
-                continue;
-            }
-
-            // Cache hit: return pre-computed result without executing the tool.
-            if let Some(cached_output) = cache_hits[idx].clone() {
-                tracing::debug!(
-                    tool = %tc.name,
-                    "[tool-cache] returning cached result, skipping execution"
-                );
-                tier_futs.push((idx, Box::pin(std::future::ready(Ok(Some(cached_output))))));
-                continue;
-            }
-
-            // Rate limiter: check the pre-computed batch result for this call.
-            if let Some(ref exceeded) = rate_results[tier_local_idx] {
-                tracing::warn!(
-                    tool = %tc.name,
-                    category = exceeded.category.as_str(),
-                    limit = exceeded.limit,
-                    "tool rate limiter: blocking call"
-                );
-                self.update_metrics(|m| m.rate_limit_trips += 1);
-                self.push_security_event(
-                    zeph_common::SecurityEventCategory::RateLimit,
-                    tc.name.as_str(),
-                    format!(
-                        "{} calls exceeded {}/min",
-                        exceeded.category.as_str(),
-                        exceeded.limit
-                    ),
-                );
-                tier_futs.push(ready_fut(
-                    idx,
-                    skipped_output(tc.name.clone(), exceeded.to_error_message()),
-                ));
-                continue;
-            }
-
-            if let Some(fut) = self.run_before_tool_hooks(idx, tc, call).await {
-                tier_futs.push(fut);
-                continue;
-            }
-
-            // Speculative try_commit (#3641): reuse a pre-executed result when available.
-            // Uses the LLM-assigned `tool_use_id` (tc.id) for result routing (critic H3).
-            // TODO(#3645): add circuit-breaker check when implemented.
-            if let Some(engine) = self.services.speculation_engine.as_ref()
-                && let Some(result) =
-                    crate::agent::speculative::stream_drainer::try_commit_with_timeout(engine, call)
-                        .await
-            {
-                tracing::debug!(tool = %tc.name, llm_id = %tc.id, "speculative try_commit hit");
-                tier_futs.push((idx, Box::pin(std::future::ready(result))));
-                continue;
-            }
-
-            tier_futs.push(self.make_exec_future(idx, tc, call, semaphore));
-        }
-        Ok(tier_futs)
-    }
-
-    fn make_exec_future(
-        &self,
-        idx: usize,
-        tc: &zeph_llm::provider::ToolUseRequest,
-        call: &ToolCall,
-        semaphore: &std::sync::Arc<tokio::sync::Semaphore>,
-    ) -> (usize, ToolExecFut) {
-        let sem = std::sync::Arc::clone(semaphore);
-        let executor = std::sync::Arc::clone(&self.tool_executor);
-        let call = call.clone();
-        let tool_name = tc.name.clone();
-        let tool_id = tc.id.clone();
-        let fut = async move {
-            let _permit = sem.acquire().await.map_err(|_| {
-                zeph_tools::ToolError::Execution(std::io::Error::other(
-                    "semaphore closed during tool execution",
-                ))
-            })?;
-            executor
-                .execute_tool_call_erased(&call)
-                .instrument(tracing::info_span!(
-                    "tool_exec",
-                    tool_name = %tool_name,
-                    idx = %tool_id
-                ))
-                .await
-        };
-        (idx, Box::pin(fut))
-    }
-
-    /// Poll tier futures concurrently with cancellation and MCP elicitation drain.
-    ///
-    /// Returns `Ok(None)` when the user cancelled (the tier loop must `return Ok(None)`),
-    /// or `Ok(Some(results))` on completion.
-    async fn execute_tier_join(
-        &mut self,
-        futs: Vec<ToolExecFut>,
-        cancel: &tokio_util::sync::CancellationToken,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-    ) -> Result<
-        Option<Vec<Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>>>,
-        crate::agent::error::AgentError,
-    > {
-        let mut join_fut = std::pin::pin!(futures::future::join_all(futs));
-        // Take elicitation_rx out of self so we can hold &mut self for handling.
-        let mut elicitation_rx = self.services.mcp.elicitation_rx.take();
-        let result = loop {
-            tokio::select! {
-                results = &mut join_fut => break results,
-                () = cancel.cancelled() => {
-                    self.services.mcp.elicitation_rx = elicitation_rx;
-                    self.tool_executor.set_skill_env(None);
-                    tracing::info!("tool execution cancelled by user");
-                    self.update_metrics(|m| m.cancellations += 1);
-                    self.channel.send("[Cancelled]").await?;
-                    self.persist_cancelled_tool_results(tool_calls).await;
-                    return Ok(None);
-                }
-                event = recv_elicitation(&mut elicitation_rx) => {
-                    if let Some(ev) = event {
-                        self.handle_elicitation_event(ev).await;
-                    } else {
-                        tracing::debug!("elicitation channel closed during tier exec");
-                        elicitation_rx = None;
-                    }
-                }
-            }
-        };
-        self.services.mcp.elicitation_rx = elicitation_rx;
-        Ok(Some(result))
-    }
-
-    /// Store tier execution results, update the dependency-failure set, prime the result cache,
-    /// update the dependency graph, and run `RuntimeLayer::after_tool` hooks.
-    #[allow(clippy::too_many_arguments)]
-    async fn apply_tier_results(
-        &mut self,
-        indices: Vec<usize>,
-        tier_results: Vec<Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>>,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-        calls: &[ToolCall],
-        cache_hits: &[Option<zeph_tools::ToolOutput>],
-        args_hashes: &[u64],
-        failed_ids: &mut std::collections::HashSet<String>,
-        tool_results: &mut [Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>],
-    ) {
-        for (idx, result) in indices.into_iter().zip(tier_results) {
-            // IMP-02: Err(_) covers all error variants including ConfirmationRequired.
-            // Ok(Some(out)) with "[error]" prefix covers synthetic/blocked results.
-            let is_failed = match &result {
-                Err(_) => true,
-                Ok(Some(out)) => out.summary.starts_with("[error]"),
-                Ok(None) => false,
-            };
-            if is_failed {
-                failed_ids.insert(tool_calls[idx].id.clone());
-            }
-
-            // Store successful, non-cached results in the tool result cache.
-            if !is_failed
-                && cache_hits[idx].is_none()
-                && zeph_tools::is_cacheable(tool_calls[idx].name.as_str())
-                && let Ok(Some(ref out)) = result
-            {
-                let key =
-                    zeph_tools::CacheKey::new(tool_calls[idx].name.to_string(), args_hashes[idx]);
-                self.tool_orchestrator.result_cache.put(key, out.clone());
-            }
-
-            // Record successful tool completions for the dependency graph (#2024).
-            if !is_failed && self.services.tool_state.dependency_graph.is_some() {
-                self.services
-                    .tool_state
-                    .completed_tool_ids
-                    .insert(tool_calls[idx].name.to_string());
-            }
-
-            // RuntimeLayer after_tool hooks.
-            if !self.runtime.config.layers.is_empty() {
-                let conv_id_str = self
-                    .services
-                    .memory
-                    .persistence
-                    .conversation_id
-                    .map(|id| id.0.to_string());
-                let ctx = crate::runtime_layer::LayerContext {
-                    conversation_id: conv_id_str.as_deref(),
-                    turn_number: u32::try_from(self.services.sidequest.turn_counter)
-                        .unwrap_or(u32::MAX),
-                };
-                for layer in &self.runtime.config.layers {
-                    let hook_result =
-                        std::panic::AssertUnwindSafe(layer.after_tool(&ctx, &calls[idx], &result))
-                            .catch_unwind()
-                            .await;
-                    if hook_result.is_err() {
-                        tracing::warn!("RuntimeLayer::after_tool panicked, continuing");
-                    }
-                }
-            }
-
-            tool_results[idx] = result;
-        }
-    }
-
-    /// Run the causal IPI post-probe after tool batch execution.
-    ///
-    /// Collects sanitized snippets from `result_parts`, runs the behavioral post-probe,
-    /// and logs a security event if the deviation exceeds the analyzer threshold.
-    /// No-op when `causal_pre_response` is `None` or no causal analyzer is configured.
-    async fn run_causal_ipi_post_probe(
-        &mut self,
-        causal_pre_response: Option<(String, String)>,
-        result_parts: &[MessagePart],
-    ) {
-        let Some((pre_response, context_summary)) = causal_pre_response else {
-            return;
-        };
-        let snippets: Vec<String> = result_parts
-            .iter()
-            .filter_map(|p| {
-                if let MessagePart::ToolResult {
-                    content, is_error, ..
-                } = p
-                {
-                    if *is_error {
-                        Some(zeph_sanitizer::causal_ipi::format_error_snippet(content))
-                    } else {
-                        Some(zeph_sanitizer::causal_ipi::format_tool_snippet(content))
-                    }
-                } else {
-                    None
-                }
-            })
-            .collect();
-        let tool_snippets = if snippets.is_empty() {
-            "[empty]".to_owned()
-        } else {
-            snippets.join("---")
-        };
-        let Some(ref analyzer) = self.services.security.causal_analyzer else {
-            return;
-        };
-        match analyzer.post_probe(&context_summary, &tool_snippets).await {
-            Ok(post_response) => {
-                let analysis = analyzer.analyze(&pre_response, &post_response);
-                if analysis.is_flagged {
-                    let pre_excerpt = &pre_response[..pre_response.floor_char_boundary(100)];
-                    let post_excerpt = &post_response[..post_response.floor_char_boundary(100)];
-                    tracing::warn!(
-                        deviation_score = analysis.deviation_score,
-                        threshold = analyzer.threshold(),
-                        pre = %pre_excerpt,
-                        post = %post_excerpt,
-                        "causal IPI: behavioral deviation detected at tool-return boundary"
-                    );
-                    self.update_metrics(|m| m.causal_ipi_flags += 1);
-                    self.push_security_event(
-                        zeph_common::SecurityEventCategory::CausalIpiFlag,
-                        "tool_batch",
-                        format!("deviation={:.3}", analysis.deviation_score),
-                    );
-                }
-            }
-            Err(e) => {
-                tracing::warn!(error = %e, "causal IPI post-probe failed, skipping analysis");
-            }
-        }
-    }
-
-    /// Process tool results after all tiers have completed.
-    ///
-    /// Clears skill env, syncs cache counters, processes each result (metrics, channel
-    /// display, VIGIL gate, sanitization, skill outcomes), flushes outcomes, runs causal
-    /// post-probe, persists the `User{ToolResults}` message, fires deferred self-reflection,
-    /// runs LSP hooks, and checks for cwd changes.
-    ///
-    /// Separated from `handle_native_tool_calls` to keep that function under the line limit.
-    #[allow(clippy::too_many_arguments)]
-    async fn process_tool_result_batch(
-        &mut self,
-        tool_calls: &[zeph_llm::provider::ToolUseRequest],
-        tool_call_ids: &[String],
-        tool_started_ats: &[std::time::Instant],
-        mut tool_results: Vec<Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>>,
-        causal_pre_response: Option<(String, String)>,
-        pending_focus_checkpoint: Option<zeph_llm::provider::Message>,
-        pending_system_hints: Vec<String>,
-    ) -> Result<(), crate::agent::error::AgentError> {
-        self.tool_executor.set_skill_env(None);
-
-        // Sync cache counters to metrics after all tool execution is complete.
-        {
-            let hits = self.tool_orchestrator.result_cache.hits();
-            let misses = self.tool_orchestrator.result_cache.misses();
-            let entries = self.tool_orchestrator.result_cache.len();
-            self.update_metrics(|m| {
-                m.tool_cache_hits = hits;
-                m.tool_cache_misses = misses;
-                m.tool_cache_entries = entries;
-            });
-        }
-
-        // Collect (name, params, output) for LSP hooks. Built during the results loop below.
-        let mut lsp_tool_calls: Vec<(String, serde_json::Value, String)> = Vec::new();
-
-        // Process results sequentially (metrics, channel sends, message parts).
-        // self_reflection is deferred until after all result_parts are assembled and user_msg
-        // is pushed to history. Calling it inside the loop would insert a reflection dialogue
-        // (User{prompt} + Assistant{response}) between Assistant{ToolUse} and User{ToolResults},
-        // violating the OpenAI/Claude API message ordering protocol → HTTP 400.
-        let mut result_parts: Vec<MessagePart> = Vec::new();
-        // Accumulates injection flags across all tools in the batch (Bug #1490 fix).
-        let mut has_any_injection_flags = false;
-        // Deferred self-reflection: set to the sanitized error output of the first failing tool
-        // that is eligible for reflection. Consumed after user_msg is pushed to history.
-        let mut pending_reflection: Option<String> = None;
-        // Accumulate skill outcomes during the tool loop; flushed once after the loop via
-        // flush_skill_outcomes to avoid N×M×13 sequential SQLite awaits (#2770).
-        let mut pending_outcomes: Vec<crate::agent::learning::PendingSkillOutcome> = Vec::new();
-        for idx in 0..tool_calls.len() {
-            let tc = &tool_calls[idx];
-            let tool_call_id = &tool_call_ids[idx];
-            let started_at = &tool_started_ats[idx];
-            let tool_result = std::mem::replace(&mut tool_results[idx], Ok(None));
-            self.process_one_tool_result(
-                tc,
-                tool_call_id,
-                started_at,
-                tool_result,
-                &mut result_parts,
-                &mut lsp_tool_calls,
-                &mut has_any_injection_flags,
-                &mut pending_reflection,
-                &mut pending_outcomes,
-            )
-            .await?;
-        }
-
-        // Flush all accumulated skill outcomes from the tool batch in a single pass.
-        // This replaces the per-tool record_skill_outcomes calls that caused N×M sequential
-        // SQLite awaits (#2770).
-        self.flush_skill_outcomes(pending_outcomes).await;
-
-        self.run_causal_ipi_post_probe(causal_pre_response, &result_parts)
-            .await;
-
-        let user_msg = Message::from_parts(Role::User, result_parts);
-        // flagged_urls accumulates across ALL tools in this batch (cross-tool trust boundary).
-        // A URL from tool N's output can flag tool M's arguments even if tool M returned clean
-        // output. has_any_injection_flags covers pure text injections (no URL); flagged_urls
-        // covers URL-based exfiltration. Both are OR-combined for conservative guarding.
-        // Individual per-tool granularity would require separate persist_message calls per
-        // result, which would change message history structure.
-        let tool_results_have_flags =
-            has_any_injection_flags || !self.services.security.flagged_urls.is_empty();
-        tracing::debug!("tool_batch: calling persist_message for tool results");
-        self.persist_message(
-            Role::User,
-            &user_msg.content,
-            &user_msg.parts,
-            tool_results_have_flags,
-        )
-        .await;
-        tracing::debug!("tool_batch: persist_message done, pushing message");
-        self.push_message(user_msg);
-        tracing::debug!("tool_batch: message pushed, starting LSP hooks");
-        if let (Some(id), Some(last)) = (
-            self.msg.last_persisted_message_id,
-            self.msg.messages.last_mut(),
-        ) {
-            last.metadata.db_id = Some(id);
-        }
-
-        // Flush deferred start_focus checkpoint AFTER User(tool_results) so the ordering
-        // Assistant→User→System is valid for OpenAI (#3262).
-        if let Some(checkpoint) = pending_focus_checkpoint {
-            self.push_message(checkpoint);
-        }
-
-        // Flush deferred utility gate hints (Retrieve/Verify). Pushed after User(tool_results)
-        // so the ordering Assistant→User→System is valid for OpenAI (#2615).
-        for hint in pending_system_hints {
-            self.push_message(zeph_llm::provider::Message::from_legacy(
-                zeph_llm::provider::Role::System,
-                &hint,
-            ));
-        }
-
-        // Deferred self-reflection: user_msg is now in history so the reflection dialogue
-        // (User{prompt} + Assistant{response}) appends after User{ToolResults}, preserving
-        // API message ordering. Only the first eligible error per batch triggers reflection.
-        if let Some(sanitized_out) = pending_reflection {
-            match self
-                .attempt_self_reflection(&sanitized_out, &sanitized_out)
-                .await
-            {
-                Ok(_) | Err(_) => {
-                    // Whether reflection succeeded, declined, or errored: the ToolResults are
-                    // already committed to history. Return Ok regardless so the caller continues
-                    // the tool loop normally (#2197).
-                }
-            }
-        }
-
-        // Fire LSP hooks for each completed tool call (non-blocking: diagnostics fetch
-        // is spawned in background; hover calls are awaited but short-lived).
-        // `lsp_tool_calls` collects (name, params, output) tuples built during the
-        // results loop above. They are captured into a separate Vec so we can call
-        // `&mut self.services.session.lsp_hooks` without conflicting borrows.
-        //
-        // The entire batch is capped at 30s to prevent stalls when many files are
-        // modified in one tool batch (#2750). Per the critic review, a single outer
-        // timeout is more effective than per-call timeouts because it bounds total
-        // blocking time regardless of N.
-        if self.services.session.lsp_hooks.is_some() {
-            let tc_arc = std::sync::Arc::clone(&self.runtime.metrics.token_counter);
-            let sanitizer = self.services.security.sanitizer.clone();
-            let _ = self.channel.send_status("Analyzing changes...").await;
-            // TODO: cooperative MCP cancellation — dropped futures here may leave
-            // in-flight MCP JSON-RPC requests pending until the server-side timeout.
-            let lsp_result = tokio::time::timeout(std::time::Duration::from_secs(30), async {
-                for (name, input, output) in lsp_tool_calls {
-                    if let Some(ref mut lsp) = self.services.session.lsp_hooks {
-                        lsp.after_tool(&name, &input, &output, &tc_arc, &sanitizer)
-                            .await;
-                    }
-                }
-            })
-            .await;
-            let _ = self.channel.send_status("").await;
-            if lsp_result.is_err() {
-                tracing::warn!("LSP after_tool batch timed out (30s)");
-            }
-            tracing::debug!("tool_batch: LSP hooks done");
-        }
-
-        // Defense-in-depth: check if process cwd changed during this tool batch.
-        // Normally only changes via set_working_directory; this also catches any
-        // future code path that calls set_current_dir.
-        self.check_cwd_changed().await;
-
-        Ok(())
-    }
-
-    /// Spawn an audit log entry for a non-clean VIGIL verdict on a tool output.
-    ///
-    /// No-op when the verdict is `Clean` or no audit logger is configured.
-    fn fire_vigil_audit_entry(
-        &mut self,
-        tool_name: &str,
-        vigil_outcome: Option<&super::VigilOutcome>,
-    ) {
-        let (Some(vo), Some(logger)) = (
-            vigil_outcome.filter(|v| !matches!(v, super::VigilOutcome::Clean)),
-            self.tool_orchestrator.audit_logger.as_ref(),
-        ) else {
-            return;
-        };
-        let (vigil_risk, audit_result, err_cat) = if vo.is_blocked() {
-            (
-                Some(zeph_tools::VigilRiskLevel::High),
-                zeph_tools::AuditResult::Blocked {
-                    reason: "vigil_blocked".into(),
-                },
-                "vigil_blocked",
-            )
-        } else {
-            (
-                Some(zeph_tools::VigilRiskLevel::Medium),
-                zeph_tools::AuditResult::Success,
-                "vigil_sanitized",
-            )
-        };
-        let entry = zeph_tools::AuditEntry {
-            timestamp: zeph_tools::chrono_now(),
-            tool: tool_name.to_owned().into(),
-            command: String::new(),
-            result: audit_result,
-            duration_ms: 0,
-            error_category: Some(err_cat.to_owned()),
-            error_domain: Some("security".to_owned()),
-            error_phase: None,
-            claim_source: None,
-            mcp_server_id: None,
-            injection_flagged: false,
-            embedding_anomalous: false,
-            cross_boundary_mcp_to_acp: false,
-            adversarial_policy_decision: None,
-            exit_code: None,
-            truncated: false,
-            caller_id: None,
-            policy_match: None,
-            correlation_id: None,
-            vigil_risk,
-            execution_env: None,
-            resolved_cwd: None,
-            scope_at_definition: None,
-            scope_at_dispatch: None,
-        };
-        let logger = std::sync::Arc::clone(logger);
-        self.runtime.lifecycle.supervisor.spawn(
-            crate::agent::agent_supervisor::TaskClass::Telemetry,
-            "vigil-audit-log",
-            async move { logger.log(&entry).await },
-        );
-    }
-
-    /// Spawn a fire-and-forget task to record this tool call's outcome in the experience store.
-    ///
-    /// No-op when experience memory is not configured or no active conversation ID exists.
-    fn record_tool_experience(
-        &mut self,
-        tool_name: &str,
-        vigil_blocked: bool,
-        is_error: bool,
-        tool_succeeded: bool,
-        tool_err_category: Option<&zeph_tools::error_taxonomy::ToolErrorCategory>,
-        llm_content: &str,
-    ) {
-        let Some(memory) = self.services.memory.persistence.memory.as_ref() else {
-            return;
-        };
-        let Some(experience) = memory.experience.as_ref() else {
-            return;
-        };
-        let Some(conversation_id) = self.services.memory.persistence.conversation_id else {
-            return;
-        };
-        let (outcome, detail, error_ctx): (&'static str, Option<String>, Option<String>) =
-            if vigil_blocked {
-                (
-                    "blocked",
-                    Some("vigil".to_owned()),
-                    Some(truncate_utf8(llm_content, 256)),
-                )
-            } else if is_error {
-                (
-                    "error",
-                    tool_err_category.map(|c| format!("{c:?}")),
-                    Some(truncate_utf8(llm_content, 256)),
-                )
-            } else if tool_succeeded {
-                ("success", None, None)
-            } else {
-                ("unknown", None, None)
-            };
-        let exp = std::sync::Arc::clone(experience);
-        let session_id = conversation_id.0.to_string();
-        let turn = i64::try_from(self.services.sidequest.turn_counter).unwrap_or(i64::MAX);
-        let tool_name_owned = tool_name.to_owned();
-        let accepted = self.runtime.lifecycle.supervisor.spawn(
-            crate::agent::agent_supervisor::TaskClass::Telemetry,
-            "experience-record",
-            async move {
-                if let Err(e) = exp
-                    .record_tool_outcome(
-                        &session_id,
-                        turn,
-                        &tool_name_owned,
-                        outcome,
-                        detail.as_deref(),
-                        error_ctx.as_deref(),
-                    )
-                    .await
-                {
-                    tracing::warn!(
-                        tool = %tool_name_owned, outcome = %outcome, error = %e,
-                        "experience: record_tool_outcome failed",
-                    );
-                }
-            },
-        );
-        if !accepted {
-            tracing::warn!(
-                tool = %tool_name, outcome = %outcome,
-                "experience-record dropped (telemetry class at capacity)",
-            );
-        }
-    }
-
-    /// Record histogram and trace-collector telemetry for a completed tool call.
-    ///
-    /// No-op when no histogram recorder or trace collector is configured.
-    fn record_tool_execution_telemetry(
-        &mut self,
-        tool_name: &str,
-        started_at: &std::time::Instant,
-        is_error: bool,
-        output: &str,
-    ) {
-        if let Some(ref recorder) = self.runtime.metrics.histogram_recorder {
-            recorder.observe_tool_execution(started_at.elapsed());
-        }
-        if let Some(ref mut trace_coll) = self.runtime.debug.trace_collector
-            && let Some(iter_span_id) = self.runtime.debug.current_iteration_span_id
-        {
-            let latency = u64::try_from(started_at.elapsed().as_millis()).unwrap_or(u64::MAX);
-            let guard = trace_coll.begin_tool_call_at(tool_name, iter_span_id, started_at);
-            let error_kind = is_error.then(|| output.chars().take(200).collect::<String>());
-            trace_coll.end_tool_call(
-                guard,
-                tool_name,
-                crate::debug_dump::trace::ToolAttributes {
-                    latency_ms: latency,
-                    is_error,
-                    error_kind,
-                },
-            );
-        }
-    }
-
-    /// Record failure-path skill outcomes and set up deferred self-reflection.
-    ///
-    /// Returns `true` if the tool succeeded (no `[error]` / `[exit code]` in output),
-    /// `false` if the failure path was taken. `tool_err_category` is consumed via `take()`
-    /// so subsequent callers see `None` after this call.
-    fn handle_tool_failure_outcomes(
-        &mut self,
-        output: &str,
-        tool_err_category: &mut Option<zeph_tools::error_taxonomy::ToolErrorCategory>,
-        is_quality_failure: bool,
-        pending_outcomes: &mut Vec<crate::agent::learning::PendingSkillOutcome>,
-        pending_reflection: &mut Option<String>,
-    ) -> bool {
-        if output.contains("[error]") || output.contains("[exit code") {
-            let kind = tool_err_category
-                .take()
-                .map_or_else(|| FailureKind::from_error(output), FailureKind::from);
-            pending_outcomes.push(crate::agent::learning::PendingSkillOutcome {
-                outcome: "tool_failure".into(),
-                error_context: Some(output.to_owned()),
-                outcome_detail: Some(kind.as_str().into()),
-            });
-            if is_quality_failure {
-                self.provider
-                    .record_quality_outcome(self.provider.name(), false);
-            }
-            if pending_reflection.is_none()
-                && !self.services.learning_engine.was_reflection_used()
-                && is_quality_failure
-            {
-                let sanitized_out = self
-                    .services
-                    .security
-                    .sanitizer
-                    .sanitize(output, ContentSource::new(ContentSourceKind::ToolResult))
-                    .body;
-                *pending_reflection = Some(sanitized_out);
-            }
-            false
-        } else {
-            true
-        }
-    }
-
-    /// Classify a raw tool execution result into structured outcome fields.
-    ///
-    /// Extracts the output string, error flag, display metadata (diff, filter stats, locations),
-    /// and outcome classification (anomaly, quality failure, error category) from the `Result`.
-    /// Side effects: records filter metrics for successful outputs and fires security events for
-    /// invalid memory writes.
-    fn classify_tool_result(
-        &mut self,
-        tc: &zeph_llm::provider::ToolUseRequest,
-        tool_result: Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>,
-    ) -> ToolResultClassification {
-        match tool_result {
-            Ok(Some(out)) => {
-                let anomaly_outcome =
-                    if out.summary.contains("[error]") || out.summary.contains("[stderr]") {
-                        AnomalyOutcome::Error
-                    } else {
-                        AnomalyOutcome::Success
-                    };
-                if let Some(ref fs) = out.filter_stats {
-                    self.record_filter_metrics(fs);
-                }
-                let inline_stats = out.filter_stats.as_ref().and_then(|fs| {
-                    (fs.filtered_chars < fs.raw_chars).then(|| fs.format_inline(tc.name.as_str()))
-                });
-                let kept_lines = out
-                    .filter_stats
-                    .as_ref()
-                    .and_then(|fs| (!fs.kept_lines.is_empty()).then(|| fs.kept_lines.clone()));
-                ToolResultClassification {
-                    output: out.summary,
-                    is_error: false,
-                    diff: out.diff,
-                    inline_stats,
-                    kept_lines,
-                    locations: out.locations,
-                    anomaly_outcome,
-                    is_quality_failure: false,
-                    tool_err_category: None,
-                }
-            }
-            Ok(None) => ToolResultClassification {
-                output: "(no output)".to_owned(),
-                is_error: false,
-                diff: None,
-                inline_stats: None,
-                kept_lines: None,
-                locations: None,
-                anomaly_outcome: AnomalyOutcome::Success,
-                is_quality_failure: false,
-                tool_err_category: None,
-            },
-            Err(ref e) => {
-                let category = e.category();
-                let is_quality_failure = category.is_quality_failure();
-                let anomaly_outcome = if matches!(e, zeph_tools::ToolError::Blocked { .. }) {
-                    AnomalyOutcome::Blocked
-                } else if is_quality_failure && zeph_tools::is_reasoning_model(self.provider.name())
-                {
-                    AnomalyOutcome::ReasoningQualityFailure {
-                        model: self.provider.name().to_owned(),
-                        tool: tc.name.to_string(),
-                    }
-                } else {
-                    AnomalyOutcome::Error
-                };
-                if let Some(ref d) = self.runtime.debug.debug_dumper {
-                    d.dump_tool_error(tc.name.as_str(), e);
-                }
-                if tc.name == "memory_save"
-                    && matches!(e, zeph_tools::ToolError::InvalidParams { .. })
-                    && e.to_string().contains("memory write rejected")
-                {
-                    self.update_metrics(|m| m.memory_validation_failures += 1);
-                    self.push_security_event(
-                        zeph_common::SecurityEventCategory::MemoryValidation,
-                        "memory_save",
-                        e.to_string(),
-                    );
-                }
-                let feedback = zeph_tools::ToolErrorFeedback {
-                    category,
-                    message: e.to_string(),
-                    retryable: category.is_retryable(),
-                };
-                ToolResultClassification {
-                    output: feedback.format_for_llm(),
-                    is_error: true,
-                    diff: None,
-                    inline_stats: None,
-                    kept_lines: None,
-                    locations: None,
-                    anomaly_outcome,
-                    is_quality_failure,
-                    tool_err_category: Some(category),
-                }
-            }
-        }
-    }
-
-    /// Process the result of a single tool execution within the batch result loop.
-    ///
-    /// Handles metrics, trace spans, skill learning, channel display, VIGIL gate, sanitization,
-    /// experience memory recording, and appends a `MessagePart::ToolResult` to `result_parts`.
-    /// Returns `Err` only on hard channel errors (e.g., broken pipe on `send_tool_output`).
-    #[allow(clippy::too_many_arguments)]
-    async fn process_one_tool_result(
-        &mut self,
-        tc: &zeph_llm::provider::ToolUseRequest,
-        tool_call_id: &str,
-        started_at: &std::time::Instant,
-        tool_result: Result<Option<zeph_tools::ToolOutput>, zeph_tools::ToolError>,
-        result_parts: &mut Vec<MessagePart>,
-        lsp_tool_calls: &mut Vec<(String, serde_json::Value, String)>,
-        has_any_injection_flags: &mut bool,
-        pending_reflection: &mut Option<String>,
-        pending_outcomes: &mut Vec<crate::agent::learning::PendingSkillOutcome>,
-    ) -> Result<(), crate::agent::error::AgentError> {
-        let ToolResultClassification {
-            output,
-            mut is_error,
-            diff,
-            inline_stats,
-            kept_lines,
-            locations,
-            anomaly_outcome,
-            is_quality_failure,
-            mut tool_err_category,
-        } = self.classify_tool_result(tc, tool_result);
-
-        self.record_tool_execution_telemetry(tc.name.as_str(), started_at, is_error, &output);
-
-        let tool_succeeded = self.handle_tool_failure_outcomes(
-            &output,
-            &mut tool_err_category,
-            is_quality_failure,
-            pending_outcomes,
-            pending_reflection,
-        );
-        let _ = self.record_anomaly_outcome(anomaly_outcome).await;
-
-        let processed = if tc.name == OverflowToolExecutor::TOOL_NAME {
-            output.clone()
-        } else {
-            self.maybe_summarize_tool_output(&output).await
-        };
-        let body = if let Some(ref stats) = inline_stats {
-            format!("{stats}\n{processed}")
-        } else {
-            processed.clone()
-        };
-        let body_display = self.maybe_redact(&body);
-        self.channel
-            .send_tool_output(ToolOutputEvent {
-                tool_name: tc.name.clone(),
-                display: body_display.into_owned(),
-                diff,
-                filter_stats: inline_stats,
-                kept_lines,
-                locations,
-                tool_call_id: tool_call_id.to_owned(),
-                is_error,
-                terminal_id: None,
-                parent_tool_use_id: self.services.session.parent_tool_use_id.clone(),
-                raw_response: None,
-                started_at: Some(*started_at),
-            })
-            .await?;
-
-        let (processed, vigil_outcome) = self.run_vigil_gate(tc.name.as_str(), processed);
-        self.fire_vigil_audit_entry(tc.name.as_str(), vigil_outcome.as_ref());
-
-        let (llm_content, tool_had_injection_flags) = match &vigil_outcome {
-            Some(super::VigilOutcome::Blocked { sentinel, .. }) => {
-                is_error = true;
-                (sentinel.clone(), false)
-            }
-            _ => {
-                self.sanitize_tool_output(&processed, tc.name.as_str())
-                    .await
-            }
-        };
-        *has_any_injection_flags |= tool_had_injection_flags;
-
-        let vigil_blocked = vigil_outcome
-            .as_ref()
-            .is_some_and(super::VigilOutcome::is_blocked);
-        if !is_error && !vigil_blocked {
-            lsp_tool_calls.push((tc.name.to_string(), tc.input.clone(), llm_content.clone()));
-        }
-
-        if vigil_blocked {
-            pending_outcomes.push(crate::agent::learning::PendingSkillOutcome {
-                outcome: FailureKind::SecurityBlocked.as_str().into(),
-                error_context: Some("VIGIL blocked tool output".into()),
-                outcome_detail: None,
-            });
-        } else if tool_succeeded {
-            pending_outcomes.push(crate::agent::learning::PendingSkillOutcome {
-                outcome: "success".into(),
-                error_context: None,
-                outcome_detail: None,
-            });
-            self.provider
-                .record_quality_outcome(self.provider.name(), true);
-        }
-
-        self.record_tool_experience(
-            tc.name.as_str(),
-            vigil_blocked,
-            is_error,
-            tool_succeeded,
-            tool_err_category.as_ref(),
-            &llm_content,
-        );
-
-        // PASTE: record tool transition for pattern learning (#3642).
-        self.observe_paste_transition(tc, started_at, tool_succeeded, vigil_blocked)
-            .await;
-
-        result_parts.push(MessagePart::ToolResult {
-            tool_use_id: tc.id.clone(),
-            content: llm_content,
-            is_error,
-        });
-        Ok(())
-    }
-
-    /// Handle a focus tool call (`start_focus` / `complete_focus`) directly on the Agent (#1850).
-    ///
-    /// Returns the tool result string. On error the string begins with `[error]`.
-    ///
-    /// ## S4 fix
-    ///
-    /// If `complete_focus` is called without an active focus session, or the checkpoint marker
-    /// is not found in the message history, an `[error]` result is returned to the LLM so it
-    /// knows the state is invalid rather than silently succeeding.
     pub(crate) fn handle_focus_tool(
         &mut self,
         tool_name: &str,
@@ -3805,6 +950,7 @@ impl<C: Channel> Agent<C> {
     /// Guards:
     /// - Returns error if a focus session is active (would interfere with focus boundaries).
     /// - Returns error if a compression is already in progress (concurrency guard).
+    #[tracing::instrument(name = "core.tool.handle_compress_context", skip_all, level = "debug")]
     pub(crate) async fn handle_compress_context(&mut self) -> String {
         use zeph_llm::provider::LlmProvider as _;
 
@@ -3941,6 +1087,11 @@ impl<C: Channel> Agent<C> {
     /// Called on early-return cancellation paths where the assistant `ToolUse` message was already
     /// persisted but the matching user `ToolResult` message was not yet written. Without this, the
     /// DB contains an orphaned `ToolUse` that will trigger a Claude API 400 on the next session.
+    #[tracing::instrument(
+        name = "core.tool.persist_cancelled_tool_results",
+        skip_all,
+        level = "debug"
+    )]
     pub(crate) async fn persist_cancelled_tool_results(
         &mut self,
         tool_calls: &[zeph_llm::provider::ToolUseRequest],
@@ -3957,74 +1108,6 @@ impl<C: Channel> Agent<C> {
         self.persist_message(Role::User, &user_msg.content, &user_msg.parts, false)
             .await;
         self.push_message(user_msg);
-    }
-
-    /// Record a PASTE pattern transition after a tool call completes (#3642).
-    ///
-    /// Resolves the owning skill via the `tool_to_skill` map built at skill activation,
-    /// then calls [`PatternStore::observe`] with the previous tool name and the outcome.
-    /// All errors and timeouts are silently ignored — PASTE failures must never block the agent.
-    #[tracing::instrument(name = "core.tool.observe_paste", skip_all, fields(tool = %tc.name))]
-    async fn observe_paste_transition(
-        &mut self,
-        tc: &zeph_llm::provider::ToolUseRequest,
-        started_at: &std::time::Instant,
-        tool_succeeded: bool,
-        vigil_blocked: bool,
-    ) {
-        let Some(ref store) = self.services.tool_state.pattern_store.clone() else {
-            return;
-        };
-
-        let tool_name = tc.name.as_str();
-
-        let Some((skill_name, skill_hash)) = self
-            .services
-            .tool_state
-            .tool_to_skill
-            .get(tool_name)
-            .cloned()
-        else {
-            return;
-        };
-
-        let prev_tool = self
-            .services
-            .tool_state
-            .last_tool_per_skill
-            .get(&skill_name)
-            .cloned();
-
-        let outcome = if tool_succeeded && !vigil_blocked {
-            crate::agent::speculative::paste::ToolOutcome::Success
-        } else {
-            crate::agent::speculative::paste::ToolOutcome::Failure
-        };
-
-        let latency_ms = u64::try_from(started_at.elapsed().as_millis()).unwrap_or(u64::MAX);
-
-        let args_json = serde_json::to_string(&tc.input).unwrap_or_default();
-
-        let _ = tokio::time::timeout(
-            std::time::Duration::from_millis(200),
-            store.observe(
-                &skill_name,
-                &skill_hash,
-                prev_tool.as_deref(),
-                tool_name,
-                &args_json,
-                outcome,
-                latency_ms,
-            ),
-        )
-        .await;
-
-        // Update last_tool_per_skill for this skill so the next tool in the same turn
-        // uses the correct prev_tool value.
-        self.services
-            .tool_state
-            .last_tool_per_skill
-            .insert(skill_name, tool_name.to_owned());
     }
 }
 
@@ -4132,58 +1215,7 @@ fn build_gated_defs_for_iteration(
 ///
 /// Returns `None` when the receiver is absent (no MCP elicitation configured) or the channel
 /// is closed, causing the `select!` branch to be disabled rather than polling indefinitely.
-async fn recv_elicitation(
-    rx: &mut Option<tokio::sync::mpsc::Receiver<zeph_mcp::ElicitationEvent>>,
-) -> Option<zeph_mcp::ElicitationEvent> {
-    match rx {
-        Some(r) => r.recv().await,
-        None => std::future::pending().await,
-    }
-}
 
-/// Build a skipped `ToolOutput` for utility gate non-`ToolCall` actions.
-///
-/// All four non-execute arms (Respond/Retrieve/Verify/Stop) produce an identical struct shape
-/// with zero blocks executed and no diff/filter metadata.
-fn skipped_output(
-    tool_name: impl Into<zeph_common::ToolName>,
-    summary: impl Into<String>,
-) -> zeph_tools::ToolOutput {
-    zeph_tools::ToolOutput {
-        tool_name: tool_name.into(),
-        summary: summary.into(),
-        blocks_executed: 0,
-        filter_stats: None,
-        diff: None,
-        streamed: false,
-        terminal_id: None,
-        locations: None,
-        raw_response: None,
-        claim_source: None,
-    }
-}
-
-/// Wrap a pre-built `ToolOutput` as an immediately-ready boxed future suitable for
-/// insertion into the tier futures list.
-fn ready_fut(idx: usize, out: zeph_tools::ToolOutput) -> (usize, ToolExecFut) {
-    (idx, Box::pin(std::future::ready(Ok(Some(out)))))
-}
-
-/// Truncate `s` to at most `max_bytes` bytes on a valid UTF-8 char boundary.
-///
-/// Never panics: if `s` is shorter than `max_bytes`, it is returned as-is.
-fn truncate_utf8(s: &str, max_bytes: usize) -> String {
-    if s.len() <= max_bytes {
-        return s.to_owned();
-    }
-    let mut end = max_bytes;
-    while end > 0 && !s.is_char_boundary(end) {
-        end -= 1;
-    }
-    s[..end].to_owned()
-}
-
-// T-CRIT-02: handle_focus_tool tests — happy path, error paths, checkpoint pinning (S5 fix).
 #[cfg(test)]
 mod tests {
     use std::sync::Arc;
